Don't scale up small browser action icons.
When creating the canvas, take cues from the image we
just loaded and don't hard code it to 19x19. This way
we can have smaller than 19x19 icons that show up
centered within the browser icon button.

This doesn't attempt to deal with scaling down images
that are too large to fit the 19x19 canvas, but we
haven't decided how to handle those anyway.

BUG=25487
TEST=Load the api_test/browser_action_tab_specific_state
extension and click on the browser action. The red
square within the button should not grow in size as you
click.

Review URL: http://codereview.chromium.org/337026