With the change in the constructor for the ChromeThread class, we are seeing a DCHECK fire in the Chrome frame network tests for whether the current thread is the UI thread. We need to invoke the correct constructor.
The other fix is in our com_util::IDispatchImpl class where we end up crashing if we fail to find
the type information for the passed in interface. Strangely this is failing for the IChromeFrame
interface which is an IDispatch derivative and seems to be correctly registered. It also fails
with the ATL IDispatchImpl classes though. However we should not crash here :) I also removed
the NOTREACHED as it keeps firing on debug test runs. Will take a look at this closely later.

TBR=tommi

Review URL: http://codereview.chromium.org/343013