Improve memwatch tool.
I removed a leak of "stack" objects, and made them a little
more lightweight by removing an unused member.

This change is only part of an option for chrome, that is not currently
linked in during any tests.  It can only currently be used under XP, so
build-bot tests are not currently appropriate.

r=mbelshe
Review URL: http://codereview.chromium.org/345033