Attempt 2 at:
Gets find bar animation/clipping to work on views/gtk.

The only difference between this and the first version is fixing an
include in extension_host that was triggering mac to build views, as
well as an extraneous SetSlideDuration I had used for testing.

BUG=none
TEST=none
TBR=ben
Review URL: http://codereview.chromium.org/342116