Fixes bug where clicking on bookmark in a folder or home button doesn't revert the autocomplete. The current approach of sprinkling Revert throughout the code is easy to forget. I've put the code in Browser so that everyone should pick it up.
If you're ok with this, I'll see if there are some calls to Revert I
can remove.

BUG=25842
TEST=add the home button to the toolbar, navigate to a page other than
the home page, type some text in the omnibox, click the home button,
make sure the text in the omnibox changes to show the url of the home
page.

Review URL: http://codereview.chromium.org/361008