Disabled extensions should be uninstallable.
The code to display the dialog looks up an
Extension by id but doesn't return disabled ones.

So I added that option to the function. 

I also added a check so that we don't show the
Options button enabled when the extension is
disabled, since you need the extension to be
enabled for the Options page to work.

BUG=26910
TEST=Disable an extension that has an Option page,
make sure Options is not enabled after disabling,
then uninstall it. It should uninstall.

Review URL: http://codereview.chromium.org/385014