Remove the ability to use remote fonts on the Stable branch.
This makes the --enable-remote-fonts flag fail on stable, but
it's not a supported stable feature. It's better to be safe and
remove this code, since it forces a load of t2embed that's
not needed in 99.99% of cases.

TBR= jungshik,ifette,yusukes
BUG= None
TEST= Run depends.exe on chrome.dll and verify that it does not have 
an explicit dependency on t2embed.dll.

Review URL: http://codereview.chromium.org/385027