GTK: Use themed background images in toolbar buttons.
When a chrome theme asks to use a background image on a toolbar button,
use it. Oddly enough, all of the graphics drawing was written but the
plumbing of actually setting the background wasn't...

TEST=Good Smile, Chuck Anderson, etc. should have button backgrounds which
are very different from the toolbar background in the GTK port.
BUG=NONE

Review URL: http://codereview.chromium.org/384057