Remove the unused ProcessFilter destructor.
Adding some code elsewhere was causing pure call errors related to this
code, but only for Release builds on gcc 4.2.4.  My gcc 4.4.1 computer
was fine.  I know better than to claim it's a compiler bug, but this
destructor is unused anyway so this change is harmless.

Review URL: http://codereview.chromium.org/402080