Implement "Import Bookmarks and Settings".
Changed to MainMenu.xib: Added a new menu item under the ^IDS_SHORT_PRODUCT_NAME menu with the name .  This menu item has a tag of 40014 corredponsing to IDC_IMPORT_SETTINGS and is connected to the commandDispatch: action of the CrApplication (First Responder).

Changes to Toolbar.xib: Added a new menu item with the name ^IDS_IMPORT_SETTINGS in the same manner as above.

BUG=26157
TEST=Launch Chrome and inspect the current bookmarks and settings.  With no browser window open and then with a browser window open choose the "Import Bookmarks and Settings..." menu option found under the "Chromium" menu.  Also, with a browser window open, hoose the "Import Bookmarks and Settings..." menu option found under the page's tool menu.  Select a browser from the popup, choose the items to be imported, and press OK.  Then inspect the bookmarks and other settings to insure that the selected items have indeed been imported.

Review URL: http://codereview.chromium.org/378036