Remove invuca.icu table from the ICU data file again (part 2 : Windows. For part 1, see http://codereview.chromium.org/399089/show)
It was added to make Chrome not crash with a webkit patch for 
bug 30437 (https://bugs.webkit.org/show_bug.cgi?id=30437) 

However, the webkit patch was rolled back so that we don't need invuca.icu for now. Removing it will cut down the data size by 230k (before compression and in the Chrome executable on Linux/Mac and the ICU data dll on Linux). 
This has to be added back when it becomes necessary again to fix the aforementioned webkit bug. 

See also http://crbug.com/28132 (a bug about ~ 4kB of space wasted by copyright statements) 

TBR=evan
BUG=20406
TEST=NONE 


Review URL: http://codereview.chromium.org/418018