Mac: change tab change notifications to pass full TabChangeType.
This makes the Mac code better parallel that on other platforms. It will also be needed to make the "glow" animation for (unselected, title-changing) pinned tabs work. (I have a mostly-working patch which does the glow, but it's horribly hacky and to get it to completely work would either involve much more hackiness or significant refactoring. That's why the glow is not part of this CL.)

Also: prevent changes in background tabs (e.g., loading finished) from killing any current bookmark bar animation. Still to do on this front (part of issue 27693): be smarter about bookmark bar updates in the selected tab.

BUG=28154, 27693
TEST=Load lots of (slow-loading) pages in background tabs while pressing Shift-Cmd-B repeatedly; make sure changes in background (unselected) tabs don't cancel the animation.

Review URL: http://codereview.chromium.org/449056