Extension Doc Changes (no building or testable changes)
This includes three minor doc changes. (1) Conditional text for description of callback based on whether it is optional -- to be less confusing. (2) Setting the page title to "" from the static pageData-title no longer shows a leading " - ". (C) Allows for a page to specify whether it shows the sideNav (and index.html declines to do so).

TBR=kathyw,aa

Review URL: http://codereview.chromium.org/460039