Mac: remove extension shelf code.
Otherwise, I'll shortly have to update it and test other changes against the extension shelf (or let bitrot set in).

The extension shelf was removed from Chrome/Chromium. This code was left in for possible future use since doing so posed no cost. The corresponding code remains in the Windows version (for experimental purposes), but has also been removed from the Linux version. If this code is ever needed again, this patch will be the one to try to revert.

BUG=30178
TEST=Trybots, to make sure everything still compiles and unit tests run cleanly.

Review URL: http://codereview.chromium.org/495004