Speculative fix for a ChromeFrame crasher reported by go/crash. The crash occurs while tearing down the ChromeFrameAutomationClient instance, which only gets called from the destructor of the ChromeFramePlugin class which is a base class of the ChromeFrame ActiveX. When this executes the worker thread which is used for handling urlmon requests has been destroyed.
The crash occurs in the UrlmonUrlRequest::Stop function while posting a task back to an invalid message loop.
We already clean up the requests in the WM_DESTROY handler, meaning that there seems to be a scenario where
we don't receive a WM_DESTROY. It is unclear what that is.

The fix is to call Uninitialize on the ChromeFramePlugin object in the destructor of the ActiveX which ensures
that it executes while the thread is still valid. Added checks for the message loop in the relevant functions
in the UrlmonUrlRequest class.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=31557

Bug=31557

Review URL: http://codereview.chromium.org/518034