Linux: allow TLS renegotiation.
Renegotiation was disabled when we switched to using our internal
version of NSS. The default in the new versions is to prohibit
renegotiation. However, since we are a client this is rather
pointless. An attacker can easily convince us to start a new TLS
connection to a host if they wish.

http://codereview.chromium.org/501178

BUG=none
TEST=Go to a site that uses renegotiation (generally because they want to request a client-side cert).