Implemented bookmark manager context menus. BUG=31857 TEST=none
Mike Pinkerton recommended I use the cross-platform menu utilities for this; I decided not to do that yet, in this patch, for a few reasons:
(1) I'm in the midst of doing heavy refactoring of the tree/list controller classes in a local branch, so anything I do now would have to get changed a lot in that branch anyway.
(2) The Windows and GTK bookmark managers don't use them, so there's not existing common code to hook into. This means I'd have to write all that code, and presumably retrofit the other platform managers to use it. But we really want to get this patch into the upcoming (tonight?) dev-channel build to avoid user confusion.

I have not added a menu to the nib. I've added code to BookmarkManagerController to create the NSMenu programmatically. I can switch to the cross-platform utilities after my refactoring.

Another note: Yes, there is some duplication of code between the Groups and Tree controller classes. I'm not worrying about this because the Groups controller class is going away completely in my refactoring, so it only has a few more days to live. :-)

Review URL: http://codereview.chromium.org/549010