SSL: don't ask for next proto if we got an error.
GetNextProto only returns an error if the NSPR socket passed in is not a valid
SSL socket. However, this happens in tests (see the bug).

It appears that in some code paths the socket is invalid by the time that we
call HttpNetworkTransaction::DoSSLConnectComplete.

This also fixes the case where we would use SPDY if the server supports NPN,
but didn't actually advertise support for it.

BUG=31611