Rename CloseIdleConnections -> CloseCurrentConnections.
This method is (and was) just for debugging; while it was closing
idle connections, that was not aggressive enough; the benchmark needs
to close all connections, not just the idle ones.

To ensure connections are abandoned, create a new pool; leaving the old pool
to languish as any pending sockets die.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/549031