Make the fyi chrome builders predict our ability to compile for release builds.
The intent of this change is to build all the targets that a chrome official
release builder does and have these buildbots be monitored on the main waterfall
page. I did not change the windows builder to be 'official' at this time, it
would more than double the compile time due to whole program optimization,
and it is not clear that running wpo on a builder that does not have similiar
system specs as the official builders would be predictive. 

We also use a changelist scoring script to determine which CL to start a
release build with, having these builders actually run tests would be
duplicative. Later it may be interesting to have these builders run only the
tests that are consistently failing on the release builders (i.e. webkit on 
linux).

Review URL: http://codereview.chromium.org/545100