Pass more args into the gae stream.
Update the push() method to support being called with
positional varargs.  Update the dict structure we pass
to the server to an array in all cases containing the
args and each class name.  Allow strings to be encoded
natively.

Fix a call to load_queue_from_disk().

BUG=22839
TEST=build info is sent with build steps
Review URL: http://codereview.chromium.org/551155