Add overflow menu to the browser action container (part 2 of supporting resize for the container).
Also improved RTL support a bit (the divider wasn't drawn on
the right side of the container and resizing was reversed).

BUG=32101
TEST=Overflow menu for browser action container should now work.
Make sure to test also right-clicking on a menu item in the overflow
menu to bring up a context menu for that item. And resizing the
container in RTL locales should work.

Review URL: http://codereview.chromium.org/557006