Merge 37244 - Separate out pageloadtimes to distinguish reload etc. from navigation
We now have 4 categories, and we gather histogram for each
of those categories.

In addition, I transitioned all the A/B experiments to use
the data only when a "real" navigation takes place, and not
when a reload or back/forward is instigated.

BUG=33229
r=darin
Review URL: http://codereview.chromium.org/555109

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/556038