Update webkit to r53957.
This includes changes to WebKit that make history.{back,forward,go}
behave synchronously when navigating back to a reference fragment.
That is something we still implement asynchronously.  As a result,
we now fail the location-hash.html test.

Includes changes to glue_serialize.cc to support serializing the
WebHistoryItem's documentSequenceNumber.  (+sky for review on this
part.)

TBR=pfeldman,sky

Review URL: http://codereview.chromium.org/556055