Resolve issues with GlueSerializeTest.BackwardsCompatibleTest by appending the new documentSequenceNumber entry to the end of the pickle.
This test forges kVersion, but the function to serialize a
WebHistoryItem does not take the current value of kVersion
into account.  It did not do so before, so instead of trying
to make it do so now, I just chose to append the new field.

TBR=sky

Review URL: http://codereview.chromium.org/556057