Get rid of check that caused redirects to fail extension updates.
When the extension autoupdater downloads .crx files, we rely on the UrlFetcher 
class to do it for us. Inside OnCrxFetchComplete I had put a sanity check that 
the url we finished with was the same one we started with, but of course this 
breaks the case of redirects and wasn't needed because the API of UrlFetcher is
that there's one UrlFetcher per outstanding fetch.

BUG=26177
TEST=Extension Autoupdate should keep working normally, and work with redirects
such as the example mentioned in the bug.


Review URL: http://codereview.chromium.org/549215