Task manager model: Make a find quicker.
Given a list such as

00000222223333

and an index i into that list, find the first occurrence of the value of list[i]. The current code just does a plain search. The new code does a reverse search starting at i.

The reason I am bothering to optimize this is because it is used in TaskManagerGtk::CompareImpl, pending my other cl.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/553155