Wire popup blocking UI to the new machinery and port over the user's old whitelist.
Notably, this removes the UI entirely for Mac and Linux.  We need to rebuild it.

This also guts the old system's testing, since most of it disappeared or changed radically.  We should test the new stuff.  I will file a followup bug for that.

There are various tiny edge cases, like if you click the address bar icon really quickly, sometimes you'll get popups without any title yet, which makes them leave gaps in the bubble that appears.  We can fix that sort of thing.  The critical bit I tried to ensure was that we never try to open a dead popup or use a dead TabContents, no matter what.

BUG=33314
TEST=Go visit some sites with popups and play with things
Review URL: http://codereview.chromium.org/562013