Update text in Source and ShellCommand's status on hookup.
start does not set text in Source and ShellCommand's
status soon enough to be captured in the status
receiver's notification.  Setting the text here prepares
the status for being sent when start is called.

Set a default description and descriptionDone for
TreeSize to fix an error in a test.

Add a test to verify that a ShellCommand's step_status
text is correctly set at setup and after it is complete.

This change is already in upstream Buildbot.

BUG=none
TEST=Source and ShellCommand status is correctly set
when initially sent
Review URL: http://codereview.chromium.org/571016