Add IPv6 probing support, and disable IPv6 resolution when it is useless
I've added minimal probing to check if IPv6 is at all possible, and when it is not, then we disable IPv6 resolution. 

I've also added histograms and A/B test support to evaluate the impact of 
this change. 

(I landed originally, but had tree problems, and this is a new CL to tryto reland).

Note that I've switched back to MACRO style enums as well, per http://dev.chromium.org/developers/coding-style
(search for "enum").

This version now does the conditional testing at a higher level (in io_thread.h), so that it should interfere less with other testing.

r=wtc,eroman 
Review URL: http://codereview.chromium.org/585005