Add builders argument support, build steps require buildername, add projecttag.
Allow specifying builders in URL argument.

Build numbers are unique to a builder, so build
steps that reference a build number must also
specify their builder name.

Add a projecttag to the datastore.  This is filled in as
'test' for now, in a future version it will be supplied
by the gae_push module.

BUG=34084
TEST=build steps work with multiple builders, builders
can be asked for explicitly, projects are labeled with
a special tag.