Merge 37383 - Refactor extension autoupdater.
This includes two changes:

1) Stop sending the Omaha UID to the gallery in favor of a "ping" parameter
included at most once per day in a update manifest fetch, indicating the 
number of days since the last time we sent the ping parameter. The calculation
of this parameter is based on the offset from now to a value in the *previous* 
response from the server where it indicated it's notion of when that day had 
started.

2) Batch update manifest requests for extensions with the same update url. The
server and protocol have supported this for a while but this is the first time
we've added support in the client.


BUG=b/2367191
TEST=Extension updates should still work normally.

Review URL: http://codereview.chromium.org/558005

TBR=asargent@chromium.org
Review URL: http://codereview.chromium.org/600005