Initial work on making extensions work in incognito mode.
This merely adds a way to enable content scripts and browser actions in incognito windows. They still don't work properly because none of the APIs work with incognito tabs.

The way to enable an extension is to add an "incognito" bit in the user prefs file. My plan is to add UI for this later.

BUG=32365

Review URL: http://codereview.chromium.org/567037