Hack DatabaseTrackerChromium.cpp on the 249 branch.
I made changes to get this to build. Some changes had been merged in here
that depended on scriptExecutionContext. I removed those and tried to fill 
things in to how they looked on WebKit's trunk _before_ the refactoring
but after removeOpenDatabaseOnMainThread was implemented.

My goal is to get the branch building again. If this patch is no good, 
revert it and r38932.

BUG= build fail
TEST= build win
TBR= dumi@chromium.org, jochen@chromium.org

Review URL: http://codereview.chromium.org/604046