Use regular message loop for cros login window
First we have to allow the regular initialization to happen.
This will permit the login wizard screens to use the network
and rendering capabilities of chrome safely. Once a user logs in
and cros mounts their encrypted directory as a profile, we
change the default profile to that one and create a new browser
window with it as well. This ought to make launching the
first browser window even faster.

Review URL: http://codereview.chromium.org/604052