Revert r39261 on 307.
1. process_commit_response_command_unittest.cc fails to compile.
  SPECIFICS and SERVER_SPECIFICS are not defined. One example build slave:
http://chrome-master.mtv:8010/builders/chrome%20official%20linux/builds/620/steps/compile/logs/stdio

2. I tried to comment out those lines (and fix chrome_tests.gypi), but the
test fails.

3. I don't want to bring in the hairier r37253 (http://codereview.chromium.org/554066), 
which would probably fix this.

4. Therefore, I'm reverting and leaving it up to Nick to either give up or 
merge and fix the issues on 307.


-- Original CL (r39261) ---------
Merge 38472 - Fix a bug where positions inside of a newlycommitted folder would be
reversed if the children of the folder weren't committed in the same
batch as the folder.

Refactor ApplyUpdatesCommandTest, pulling out functionality that would
be useful for any SyncerCommand test, into syncer_command_test.h.

Add a test case for ProcessCommitResponseCommand using the new
SyncerCommandTest framework.  Add a test for the bug.

BUG=33081
TEST=sync_unit_tests. Also, manual testing, using the reduced repro
instructions described in comment #26 of the bug.


Review URL: http://codereview.chromium.org/652022