Add "Force Reload This Page" menu item on Mac.
This CL adds a menu item "Force Reload This Page" with keyboard shortcut Command-Shift-R to the View menu in MainMenu.xib as an alternative, meaning it is shown instead of "Reload This Page" when the shift key is held down.

BUG=36245
TEST=Open the view menu. Pressing shift should change "reload this page" into
"force reload this page", which should do what it says. Also, Command-Shift-R
should do the same thing. Astronomy picture of the day is a good test site for
this.

Review URL: http://codereview.chromium.org/651040