Landing this patch on behalf of Mark Hahnenberg <mhahnenb(at)gmail.com>
Added a vector of Time objects to the AutofillEntry object for storingthe results retrieved from the date_created field of the autofill_datestable corresponding to the AutofillKey retrieved from the autofill table.

Also modified the old GetAllAutofillEntries test to account for this newlystored data.

BUG=35230

TEST=WebDatabaseTest.GetAllAutofillEntries

Review URL: http://codereview.chromium.org/600070