Make the js console open by default when you click to inspect an extension view
This has been a long-time pet peeve of mine that it takes an extra click for
the common case when you want to see console.log() output in the inspector for
a background page.

BUG=none
TEST=Load an extension with a background page in Developer Mode. Under "Inspect
active views", click on the link. The inspector should come up with the
javascript console shown.


Review URL: http://codereview.chromium.org/661418