Merge 40684 - Remove a CHECK() which was firing.
The problem was the load state poller may call OnUploadProgress() while a redirect is in progress.
It should be safe to remove the CHECK(), since it was recently added and execution used to pass through previously in this case.

BUG=37035
TBR=wtc

Review URL: http://codereview.chromium.org/669117

TBR=eroman@chromium.org
Review URL: http://codereview.chromium.org/668116