Fix launching Clear Browsing Data dialog from the history page.
The problem here was that the modal dialog would block, while it didn't block on
windows.  Since it was launched from a standard Chrome task (the DOMUI call), it
would block all other Chrome tasks.  The callbacks from the BrowsingDataRemover
to let the window know to close was a Chrome task, and so would deadlock since
it couldn't tell the dialog to close until the dialog was closed.

I spun it off into a cocoa event using |performSelector:afterDelay:| with 0 delay which freed up
Chrome tasks to be processed again.

BUG=37976
TEST=Launch clear browsing data dialog from history and make sure you can clear
the history and it closes itself.  Also ensure it still works as expected from
menu and preferences.

Review URL: http://codereview.chromium.org/1129002