Merge WebKit 53442
When we create the document fragment from a markup string,
either to perform a paste operation or a drag and drop, we
want to remove all the event handlers and any attribute that contain
a value that leads to code execution.
The HTMLParser class is now aware of the needs of stripping these attributes.
I've modified the call to createMarkupString for every platform.

TBR= noel@google.com
BUG= b/2349715
TEST= None
Review URL: http://codereview.chromium.org/1126008