Eliminate all UI thread decoding of extension images.
Except one, that is. We have one location we need to 
take a look at (I've added a comment). This changelist 
converts UI usage of DecodeImage on the UI thread to 
a revamped and simplified ImageLoadingTracker class. 

I plan on adding to GetFilePath a DCHECK for the File 
thread but decided to do so in another changelist, 
since it has a high likelyhood of flushing something 
out and be backed out because of that. This started 
out as issue 38521 (make infobar use cached icons) 
but grew in scope to just eliminate all UI thread 
access to DecodeImage and GetFilePath. 

BUG=http://crbug.com/38521 
TEST=None (extensions should work as before)

Review URL: http://codereview.chromium.org/1075006