Override wasLastPushSuccessful in HttpStatusPush.
Track whether or last push was successful based on
which callback was issued from the connection request.
Alter popChunk to immediately use a chunk size of 1 if
the last request was not successful.  Set a default of
1MB for the max http request size, we might be able to
increase this to 9-10MB.

Fix wasLastPushSuccessful() in StatusPush.
When the last push succeeded, lastIndex is set
to queue.getIndex() in hookPushCb, so lastIndex <
queue.getIndex() would never be true.

BUG=none
TEST=the master reduces the amount of packets sent
when in a failure mode, a reasonable default is used
to measure whether large packets should be tossed
Review URL: http://codereview.chromium.org/1311002