Merge 56663 - CSSPrimitiveValue::setFloatValue/setStringValue should throw an exception if passed an incorrect unit type. <http://webkit.org/b/36680> / <rdar://problem/7801016>
Reviewed by Darin Adler.

WebCore:

Test: fast/dom/setPrimitiveValueexceptions.html

* css/CSSPrimitiveValue.cpp:
(WebCore::CSSPrimitiveValue::setFloatValue): Throw an INVALID_ACCESS_ERR if either of our type or the passedin
type is not a numeric type.
(WebCore::CSSPrimitiveValue::setStringValue): Throw an INVALID_ACCESS_ERR if either of our type or the passedin
type is not a string type.

LayoutTests:

* fast/dom/scripttests/setPrimitiveValueexceptions.js: Added.
* fast/dom/setPrimitiveValueexceptionsexpected.txt: Added.
* fast/dom/setPrimitiveValueexceptions.html: Copied from LayoutTests/fast/dom/eventattributeavailability.html.

TBR=jschuh@chromium.org
Review URL: http://codereview.chromium.org/1549001