Fix a crash when creating tabs from an Infobar by making sure the Infobar is not set as the Container when the Infobar dies or its delegate dies (so that we don't get calls from the ExtensionView after our destruction).
Some background: 

Our ExtensionInfobarDelegate owns the Infobar state, including the
ExtensionView/Host. This is because Infobars in Chrome are deleted
and recreated when you switch away from the tab and come back (and
we don't want to lose state doing so).

We also have an ExtensionInfobar object, which is set as a
"Container" for the ExtensionView, so that it can responds to
resize events from the ExtensionView.

The problem is that once chrome.tabs.create is called, a new tab
is created, which deletes the Infobar object that has yet to
receive an OnMouseLeave message from the View. Hence the crash.

We therefore null the Container in the destructor for the
ExtensionInfobar, to make sure the ExtensionView doesn't report
more events to us after we are deleted.

The only problem with this is that using the delegate_ pointer
in the destructor for ExtensionInfobar is not safe, because the
delegate may be deleted before us. Therefore, we also now observe
when the delegate is destroyed, so we can null the Container
through it.

BUG=http://crbug.com/39755
TEST=Using tabs.create should not crash the extension. Also,
opening two tabs, both with an infobar should not crash if you
switch to the first tab and close it.


Review URL: http://codereview.chromium.org/1583002