[Mac] Fix DCHECK() on page actions in preview.
Page actions can be image-less for a bit until their image loads.  The
image is visible and a default size is used.  The DCHECK() just wanted
to make sure that we weren't accidentally drawing an image of one size
into a rect of a different size.

BUG=37865
TEST=No DCHECK installing an extension.

Review URL: http://codereview.chromium.org/1605002