Refactor FileVersionInfo into an interface with platform implementations. This allows us to move the chrome specific version informaton used by Linux into src/chrome.
Add a GetChromeVersionInfo() for Linux in src/chrome/app/ and make sure to use this in src/chrome.

In src/webkit/glue, add a new glue method for getting the product version.  When compiling chrome, use an implementation in src/chrome/renderer (which uses GetChromeVersionInfo()) and a stub implementation for test_shell.

Review URL: http://codereview.chromium.org/1560027