Merging to 375 branch: http://codereview.chromium.org/1619014/show
This CL makes the translate context menu disabled when the
page language is the same as the UI language.

BUG=39506
TEST=With Chrome configure to be in English, Visit www.google.com.
Right click on the page, the context menu "Translate to English"
should be disabled. Do the same on www.google.fr, the menu should
be enabled. 
TBR=laforge

Review URL: http://codereview.chromium.org/1568038