Windows bots build and use OSMesa
This is a software GL renderer that can be used by bots that do not have access to native GL. It will be used for running tests that use GL, such as Pepper 3D, WebGL and accelerated compositing. Only working on Windows so far.

TEST=trybots
BUG=none

Review URL: http://codereview.chromium.org/1740001