Bookmark manager: partial fix position of context menu for trees.
When the user presses the context menu key we ask the tree for the rect where the context menu should be placed at. We incorrectly returned the rec for whole tree item, including its children. With this change we only return the rect for the row of the tree item.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/1758009