Merge r45628 from the trunk to the 375 branch.
Honor the select_all argument to AutocompleteEditViewMac::FocusLocation.

Calling -[NSWindow makeFirstResponder:] to give focus to the
AutocompleteTextField will select the entire contents of the field. If the
field is already focused and "select all" is not desired, avoid making this
call. If "select all" is desired, or if the field is not focused (regardless
of the value of select_all), the field will be focused and its entire contents
selected.

BUG=40322, 40682
TEST=With a profile primed so that "google.com" is the first suggestion when
     typing "g" in the location bar, and no windows open, quickly press
     command-N and begin typing "google" while the tab is being created.
     Expect all of the entered characters to make it to the location bar. When
     the bug occurs, the location bar in its entirety will be selected at some
     point during or immediately following tab creation, which will cause
     previously-typed text to be lost or the first autocomplete suggestion to
     be chosen, and something like "ogle" or "google.com/ogle" to appear in
     the location bar.
Review URL: http://codereview.chromium.org/1709010

TBR=mark@chromium.org
Review URL: http://codereview.chromium.org/1712016