Fresh ChromeFrame installs triggered via wave required IE to be restarted for wave to switch to Chrome in the current page/tab.
The ChromeFrame install flow instantiates the ChromeFrame ActiveX which ensures that the IMoniker patches are in place for switching
to work. However ChromeFrame relies on the BHO to be loaded in order to determine that a URL is a top level URL. 

The ChromeFrame ActiveX now attempts to load the BHO dynamically and registers it with the top level IWebBrowser for the page.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=42790

Bug=42790

Review URL: http://codereview.chromium.org/1789007