Disable the replacement of U+005C (backslash) with Yen sign in documents encoded in Japanese encodings. This is the minimum patch that is safe. It could have been much more extensive, but with this minimal patch, we get what we want without putting the branch at risk of unexpected regression.
This is for the 375 branch. 

BUG=9696
TEST=1. Go to any EUC-JP/Shift_JIS encoded web pages with backslash (e.g. LaTeX source, C/C++/Perl/Python source pages with character ecapes, CSS stylesheets 
with character escape, etc) and copy'n'paste a code snippet to oher programs. Backslashs should remain intact instead of being turned to Yen signs.
One webkit layout test fails as a result because the upstream has a test for this, but it shouldn't matter)
TBR=mal
Review URL: http://codereview.chromium.org/1810011