Opening new chrome frame tabs would fail randomly. This was because of a race condition between processing the connect external tab ack in chrome frame and receiving requests for the tab. As a result the new tabs would not render anything.
Fix is to resume the pending views in the OnReinitialize task in the ExternalTabContainer.

I also added a NOTREACHED when we fail to find the TabProxy for a tab handle in the chrome frame automation code to help
catch future instances of this issue.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=43376

Bug=43376

Review URL: http://codereview.chromium.org/2011006