Add --host-rules support.
The format for --host-rules is identical to --host-resolver-rules.
The difference is that --host-rules affects the endpoint of the HttpNetworkTransaction, not just the host resolver.  So, this means the host passed to the host resolver and the TCP connect(), the tunnel CONNECT, and the SOCKS connect will be different.

Review URL: http://codereview.chromium.org/2057007