Removed the "Reload" button from the HTTP cache tab of the net-internals page.
The "Reload" button was used to refresh the stats + cache listings.
With this change, I poll the statistics for updates, so the reload button is unecessary.

Note that the cache listing is too expensive to poll, so I replaced it with a link to a listings page.

Lastly, updated the text dump to include the http cache statistics.

BUG=37421

Review URL: http://codereview.chromium.org/2085005