Implement fading in for the InfoBubble base class.
For bubbles that are not shown as a direct result of
a mouse click, we use fade in. This applies to
the FirstRun bubble and the extension installed
bubble.

For bubbles that have controls (buttons/links) we
fade out if interacting with those controls closes
the bubble. This applies to the first run bubble,
app launcher, extension installed bubble, bookmark
bubble and the content settings bubble.

BUG=None
TEST=Make sure the above-mentioned bubbles fade in
and out during hide and show as described above.

Review URL: http://codereview.chromium.org/2079008