chromeos: Add _CHROME_LOGGED_IN property on root window.
This gets set to either 1 or 0 depending on whether the user
is currently logged in or not.  I'm doing this as a step towards
using the same window manager process during and after login;
this will let the WM know which log file it should be using, for
instance.

BUG=chromium-os:2790
TEST=tried it, and will add an autotest for this soon

Review URL: http://codereview.chromium.org/2078017