seccomp sandbox: Split source file list into a separate .gyp file
The intention behind this is to make it easier to sync the .gyp file
into the non-Chromium copy of the seccomp sandbox so that it can be
used to build a standalone version of the sandbox.

Also, it arguably makes the .gyp files more manageable.

Removes a dependency on "base", which the seccomp sandbox does not use.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/1939002