Add a notifier that only sends mail on regress/improve.
Create a PerfNotifier based on ChromiumNotifier that
exits early if the status box for a step doesn't mention
a regression or improvement.  If either is mentioned,
call into ChromiumNotifier's stepFinished to continue
sending the mail.

BUG=none
TEST=startup test sends mail when a regress/improve
expectation occurs, does not send mail when a regular
failure occurs.
Review URL: http://codereview.chromium.org/2428005