Fix: Clearing search 'foo', closing and reopening Find should not prepopulate with 'foo'.
We need to reset the remembered values when the user
clears the search term from the Find box to prevent
them from showing up on close+Reopen and on close+F3.

Added test to cover these cases.

This change also makes sure we don't clear the Find
string when we don't need to, which was happening on
close of the Find box and on new tab, causing the
fix not to work.

BUG=42639
TEST=Covered by new automated test.
Review URL: http://codereview.chromium.org/2322006