AutoFill Profiles dialog implemented according to new mocks on Mac
New mocks are attached to bug 44622.  These changes replace the in-place editing of address and credit card records with a table of records and separate sheets for manipulating the record data.  Changes to the layout of fields on the sheets has been done also.
AutoFillDialog.xib changes:  Replaced disclosure based list of address and credit cards with an NSTableView of the same data.  Added buttons for "Add", "Edit", and "Remove".
Replaced AutoFillAddressViewController.xib with sheet-based AutoFillAddressSheetController.xib.
Replaced AutoFillCreditCardViewController.xib with sheet-based AutoFillCreditCardSheetController.xib.

BUG=44621
TEST=AutoFillAddressModelTest,AutoFillAddressSheetControllerTest,AutoFillCreditCardModelTest,AutoFillCreditCardSheetControllerTest,AutoFillDialogControllerTest

Review URL: http://codereview.chromium.org/2673006