NaCl: Remove call to NaClAppDtor()
Calling NaClAppDtor() is unnecessary because the process is about to
exit anyway, and it happens only in the error path because
NaClAppDtor() is not able to shut down running NaCl threads safely.
There is a TODO for its removal in the NaCl tree.

This is related to http://codereview.chromium.org/2659002.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=560
TEST=nacl_ui_test

Review URL: http://codereview.chromium.org/2753002