Added BackgroundContentsService to manage lifecycle of BackgroundContents.
If --restore-background-contents flag is passed, stores the URLs of running
BackgroundContents in preferences so they can be re-launched when the browser
restarts.

Moved logic to shutdown background contents into BackgroundContentsService so
we can use this to coordinate when to keep the browser process running.

BUG=43382
TEST=new tests

Review URL: http://codereview.chromium.org/2104018