Remove pipe support from pbrpc and slavelastic.
The main driver for this is that pipe I/O on Windows cannot be async without a significant
amount of plumbing that I'm not willing to do.

Removing pipe support will remove the need for a read thread which will simplify
the code significantly and remove some dead locks.

Instead, made the unit tests test ListenTcpChannel and TcpChannel from both ways,
both from the parent and from the child processes.

Review URL: http://codereview.chromium.org/2870020