Fix BMM paste on mac and list hover artifact.
The Bookamark manager on mac would paste a url into a bookmarks url field, and
then the same paste event would propagate up to the list and paste the page as a
new bookmark.  This stops propagation of paste events from the text fields in
the list.

Also, now that the list has spacers for the dynamic rendering, the spacers would
get a 2 pixel hover effect.  This removes that.  You can see it by slowly moving
from the last bookmark down to below the bottom of the list.
BUG=43442
TEST=NONE

Review URL: http://codereview.chromium.org/2805031