Do some minor cleanup in histogram.*
I was trying to debug some code, and noticed some excess
complexity.

There are no significant semantic changes.
DCHECK on a foo.get() is a waste given we're about to deref
the pointer.
Execution cost should be be minially less, and we get rid
of several lines of code.

BUG=46263
r=mbelshe
Review URL: http://codereview.chromium.org/2805033