Push change for stevenjb:
Issue 2848016: Store wifi network in network_menu_button.cc as a path instead of as a struct.
Store wifi network in network_menu_button.cc as a path instead of as a struct. 
This should address a potential edge case where wifi network information gets updated after the menu is opened and before a network is selected. 

BUG=None 
TEST=None
Review URL: http://codereview.chromium.org/2817036