Work around exception in command "which" while running GYP inside ChromiumOS chroot.
ChromiumOS is trying to remove dpkg-buildpackage.
If running inside ChromiumOS chroot, the "which" can throw exception when the target file is missing.
( dpkg-buildpackage will be missing in this case )
This doesn't happen with the "which" on Ubuntu.

Work around it by using "whereis" instead. It works on both ChromiumOS choot and Ubuntu.

Review URL: http://codereview.chromium.org/2945011