Revert 53122 - libxml: LIBXML_STATIC isn't needed on windows either
In my previous patch today, I moved LIBXML_STATIC into a Windows-only
block in the gyp file.  It turns out I misread the #defines and it's
not needed on any OS.

libxml.h has this:
  #ifndef PIC
  #  define LIBXML_STATIC
  #endif
The PIC define mentioned there appears to be a Linuxism, which is fine
for us because we want it to be static anyway.

Review URL: http://codereview.chromium.org/2825052

TBR=evan@chromium.org
Review URL: http://codereview.chromium.org/3049004