Update Breakpad from r627 to r628 to pick up:
Separate the "hello" message from the parameter messages in
Inspector::ReadMessages as was done before r627. The "hello" message contains
the parameter count and is referenced while the message reader loops through
parameter messages. Prior to r627, both messages were named |message|, which
was confusing, probably caused a compiler warning, and apparently provided the
motivation to share them. This caused the crash inspector to fail to properly
collect the parameters. The common failure mode (although others are possible)
was for the inspector to attempt tor read more parameter messages than were
available, resulting in an IPC timeout and inspector death. No crash report
would be written, and the application expecting its crash to be inspected 
would time out waiting for a response from the inspector and then _exit. Thisis effectively a failure to properly handle crashes.

The inner message is reintroduced, and named parameter_message for
disambiguation.

BUG=49821
TEST= - Crashes catchable by the Mac Breakpad framework.
      - about:crash and its browser equivalent works and produces crash dumps
        in ~/Library/Breakpad/Chrome_Mac
Review URL: http://codereview.chromium.org/2842063