NaCl: Disable "gyp tests" step added in r53295 because it doesn't pass yet
"gyp tests" is currently failing on all of the NaCl trybots, which
renders them useless because it blocks any of the later useful steps
from running.

Change haltOnFailure to False for the Gyp steps so that problems with
the Gyp build don't stop the Scons build from running in the future.

Review URL: http://codereview.chromium.org/3007010