Revert 53332 - Update the Native Client DEPS to include r2750
This change removes a use of accept().  This will allow removal of the
special hole in the Mac sandbox that was opened for NaCl.

BUG=26781
TEST=all tests pass, especially nacl_ui_tests

This is also needed to pull in NaCl's r2744 to get sel_main_chrome.c.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=642
BUG=http://code.google.com/p/nativeclient/issues/detail?id=469

Review URL: http://codereview.chromium.org/2873059

TBR=mseaborn@chromium.org
Review URL: http://codereview.chromium.org/2847067