Remove FFmpegDemuxer::current_timestamp_ and always pass AVSEEK_FLAG_BACKWARD when seeking.
Previously we falsely assumed that AVSEEK_FLAG_BACKWARD was needed when seeking backwards in the media file.  Turns out all this flag does is seek to a point in the file that is less-than-or-equal to the desired timestamp -- regardless of the current position in the file!

Furthermore, the value of current_timestamp_ was never guaranteed to be valid as poorly muxed files tend to contain AV_NOPTS_VALUE packets.  This all resulted in non-deterministic seeking behaviour depending on whether current_timestamp_ was valid or contained AV_NOPTS_VALUE.

BUG=48722
TEST=covered by layout tests

Review URL: http://codereview.chromium.org/3050011