EncoderZlib/DecoderZlib for chromoting
Encoder and decoder using zlib for chromoting. This implementation has zero
copy out of the decoder and encoder. The consequence is that we have to break
out the zlib stream into rect boundaries which requires the synchronication
flush feature in zlib. This feature will hurt compression ratio but the effect
still need to be measured.

This patch also provides tests for testing the Encoder and Decoder pair with
zlib.

TEST=remoting_unittests

Review URL: http://codereview.chromium.org/2868062