Make the width of overlay pages flexible.
Before the change, overlays take up the whole width
of the page except padding, thus overlays are very wide.

With this change, overlays only take up the width that is
necessary to show the contents inside.

The individual overlays can control the width using CSS
as needed. If you want the original behavior (taking up
the whole width), you could add "-webkit-box-flex: 1" to
the overlay's CSS. If you want to set a fixed width, you
could add something like "width: 600px" in the overlay's
CSS.

Note that the ability to set the fixed with is necessary to
follow a suggestion from Cole regarding the size of the
overlay for adding languages. See before.png and after.png in
crosbug.com/4573 for details.

TEST=manually
BUG=chromium-os:4573

Review URL: http://codereview.chromium.org/3056042