Make reload accelerators more consistent across platforms.
On Windows and Linux (both Views and GTK+), this makes
Ctrl-R, F5, and regular clicks on the reload button do
regular reloads, while Ctrl-Shift-R, Shift-F5, Control-F5,
and holding Shift or Control while clicking the button all
bypass the cache.

It does the same for the reload button on Mac, but I'm not
modifying any Mac keyboard shortcuts in this change (it
looks like Command-Shift-R already bypasses the cache).
I've updated http://crbug.com/38356 to mention that when
the F5 shortcut is added to Mac, the cache-bypassing
modifiers should also be supported.

BUG=47383
TEST=none

Review URL: http://codereview.chromium.org/2876008