Revert 55311 - Add remoting to wrench menu.
This change adds remoting item to wrench menu.
The change is under flag enable-remoting, so end users will not see it.

The CL is cleaned up. It was compiling and working. I added the check for kEnableRemoting and resynched and it is no longer compiling.

Submitted for: pranavk@chromium.org
Reviewed: http://codereview.chromium.org/3075024/show

BUG=50242
TEST=NONE

Review URL: http://codereview.chromium.org/3080038

TBR=hclam@chromium.org