Merge 55357 - Take 2: Copy the npapi pepper test plugin during the build so we don't have to do it when running the test.
Remove all the code in the test for copying files since it's no
longer needed.

The difference from before is we only copy over the pepper test
plugin on windows because it's the only platform with tests for
it.  On Linux, unloading the plugin was causing random crashes
in tcmalloc.

Review URL: http://codereview.chromium.org/3015064

TBR=tony@chromium.org
Review URL: http://codereview.chromium.org/3105001