Support GCF as the default HTML viewer as well as supporting an exclusion list.
Now, there are two URL lists that we support and a master REG_DWORD value ("IsDefaultRenderer" ) to switch between them.
Note that the OptInUrls key is no longer supported but the URL lists mentioned below function in the same way OptInUrls used to.

This is basically how it works:

if IsDefaultRenderer
  Url list name is "RenderInHostUrls" and lists patterns that the host (IE) should render.

if not IsDefaultRenderer  (i.e. it's 0)
  Url list name is "RenderInGcfUrls" and lists patterns that GCF should render.


Also fixing typo :)

TEST=See description above and in bug report.
BUG= 50788

Review URL: http://codereview.chromium.org/3131003