net: add embedded DNSSEC chain support.
Now that the DNS root is signed we have a good trust path in several
TLDs (including .org). This patch enables self-signed certificates to
include a DNSSEC chain as an extension which proves a CERT record,
containing the fingerprint of the public key.

The format of the chain is still undecided, so this is only enabled
with --enable-dnssec-certs.

BUG=none
TEST=net_unittests

http://codereview.chromium.org/2806076