GTTF: Display information about a bot that is going to be fail earlier on the console view.
Many of our bots have multiple test steps. If any of them fails, the bot is going to turn red.
However, it's not immediately visible on the console view (the bot is still yellow).
This change makes it yellow with red border in case one of the steps failed,
but the build is still in progress.

This is based on nsylvain's patch http://codereview.chromium.org/3084022/show 

R=nsylvain
Review URL: http://codereview.chromium.org/3159011