Handle race to create histogram more cleanly
Two threads are currently allowed to attempt
to define the same histogram at the same time.
A factory class sorts out which one was first,
and ensures that only the winner remains
registered.  An old DCHECK used to assert
that only one caller (constructor) could
register a given histogram.  This CL
removes that DCHECK, and leaves the
first to register as the winner.

BUG=52394
r=eroman
Review URL: http://codereview.chromium.org/3107018