Disable autocomplete for search keywords.
We have custom UI for search keywords anyway, so in practice a user would never see the autocompleted url.  But, we had been computing the index incorrectly in some cases, as we trim off "http://" and "www." to compute the keyword.  Since we don't use it anyway, simplify the logic and get rid of it :)

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3151020