Merge 56470 - Allow data URLs to load chrome-extension:// resources. Basic notifications use data: URLs internally so this was required to support that.
Also rejigger the logic in this area again to clarify and fix
an issue where the incognito window check did not have high
enough priority.

Also add back a null-pointer check that was previously in this code and is apparently necessary with a comment.

BUG=52313,52374

Review URL: http://codereview.chromium.org/3159019

TBR=aa@chromium.org
Review URL: http://codereview.chromium.org/3184012