Landing r56534 to the M6 branch. Original review: http://codereview.chromium.org/3142015/show
Ensures we don't let users translate pages that are in an unsupported language by disabling the context menu in such cases.
It is OK to translate pages for which the language is unknown, as we let the server detect the language in such cases.

BUG=48635
TEST=Not sure how to repro without hacking the code.
You need a page in a language that the CLD detects but that Chrome does not support for translation.
TBR=kerz

Review URL: http://codereview.chromium.org/3135031