Tweak word-breaking rule for better Hebrew word breaking and fix a trivial mistranslation in Hungarian locale.
A lot of Hebrew speakers use U+0022 (double quotation mark) between
Hebrew letters instead of U+05F4 because the latter is not available
in a typical Hebrew keyboard. So, we have to treat U+0022 as "MidLetter"
when put between Hebrew letters and should not break word at U+0022 in
those cases.

This CL tweaks ICU's word.txt (rule file) for that. This was reported
to the upstream (CLDR, ICU and Unicode).

In addition, fix the translation of Vietnamese in Hungarian locale. I fixed
this in the upstream several months ago. I just forgot to pick it up the last
time I updated the ICU data.

The corresponding pre-built ICU data files for Linux, Mac, and Cr*OS 
are updated. The data file for Windows will be updated next (before 
DEPS update).

Non-Hebrew changes in word.txt is just updating word.txt file to sync with the actual data file we have been using. 


BUG=22909,34009
TEST=1. Go to http://crbug.com/22909 and type several words mentioned in the report with the spellcheck language set to Hebrew. They shouldn't be marked as misspelled. A unittest will be added when DEPS is rolled for ICU.
2. In Hungarian chrome (on Windows, start Chrome with --lang=hu. on Linux, start Chrome with 'LANGUAGE=hu chrome'), Vietnamese is translated as 'vietnami' (without any accent) in the language menu.

TBR=hbono

Review URL: http://codereview.chromium.org/3134032