Revert 57386:
The interval test started failing after we rolled back to a previous version of Webkit. This code
does not seem to support it. 

Original description: Move the chromium.Interval to chrome.Interval as part of the benchmarking
extension.  This means that users will need to use --enable-benchmarking
to access it.  It really shouldn't be part of the exposed API.



BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3126029

TBR=mbelshe@chromium.org
Review URL: http://codereview.chromium.org/3186033