Separate the "search host to TemplateURL" functionality out of the TemplateURLModel.
This separation allows for a smaller surface area which can be made read-only
thread-safe so that the question "IsSearchProviderInstalled" can be answered
from the I/O thread. The method to that answer question is not in this patch,
but this is the underlying mechanics that is necessary.

BUG=38475
TEST=unit_test --gtest_filter=Keyword* or Template* or SearchHost*

Review URL: http://codereview.chromium.org/3247003