Cleanup: make more PrefMember methods const.
This allows the users of PrefService to make their accessors
that refer to PrefMember members const too.

We use mutable for some PrefMember members, when they
do not reflect the outside-visible state. Those members
are used to do lazy loading of the preference values.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3271004