Enable webgl and accelerated compositing by default on all platforms. Removes --enable-webgl and --enable-accelerated-compositing flags and replaces them by --disable-webgl and --disable-accelerated-compositing. This CL is practically identical to: http://src.chromium.org/viewvc/chrome?view=rev&revision=58711
which was checked in the later reverted in:

http://src.chromium.org/viewvc/chrome?view=rev&revision=58730 

due to UI test failures related to video playback via the compositor. The main difference
is that mac is no longer excluded so the two features are now turned on on all 3 platforms.

The issue with the failing UI tests was corrected upstream in WebKit:
http://trac.webkit.org/changeset/66923

so this WebKit change must be rolled in before the chromium change lands.

BUG=54469
TEST=NONE


Review URL: http://codereview.chromium.org/3348012