ChromeFrame tests now use the mock http server instead of the python web server. This allows us to control the behavior of the http server in the future.
The change in the \chrome_frame_tester_helpers.js file is to send over asynchronous XHR post
requests to prevent the browser from blocking in the UI thread while waiting for a response.

We copy the the CFInstall and CFInstance scripts to the test\data folder to enable the mock
server to serve them. Added support for the /echoheader? based GET requests. Some ChromeFrame
tests rely on this.

Bug=54908

Review URL: http://codereview.chromium.org/3356013