Merge 59241 - net: disable renegotiation checks.
Although we have a handle on the problems of corporate MITM proxies,
users are now reporting the anti-virus programs are choosing to
implement their virus scanning by MITM attacking local SSL/TLS
connections.

It only appears to be some anti-virus programs in some configurations,
but we don't really have a good idea of the scope of the issue so this
patch disables the renegotiation checks (for merging to the stable
branch) while we work on a better solution.

BUG=55410
TEST=none

http://codereview.chromium.org/3340024/show

TBR=agl@chromium.org
Review URL: http://codereview.chromium.org/3315024