Revert 59718 - Merge 59102 - Change the toast text and make sure errors are written to HKLM too (not just success messages).
The toast text changes are listed in the bug; basically we
have 4 experiments that differ only in the text we display
as a headline.

I also fixed an issue raised by QA with error codes being
written to HKCU as opposed to HKLM. This should fix that
issue.

BUG=http://crbug.com/54835, http://crbug.com/44378
TEST=Requires manual testing by QA.

Review URL: http://codereview.chromium.org/3384010

TBR=kerz@chromium.org
Review URL: http://codereview.chromium.org/3427009