Fix sync_unit_tests two-phase lookup problems.
Clang fails to compile some files used by sync_unit_tests because it is more
strict on code following the Standard.

Type-dependent names are bound during template instantiation by looking at
declarations that are visible at the point of definition of the template
and declarations from namespaces associated with the types of the function
arguments both from the instantiation context and from the definition context.

For this reason, when declaring a function like
std::ostream& operator<<(std::ostream& s, const syncable::Entry& e);
it is important that the function is declared in the namespace of one of its
arguemnts, i.e. syncable in this case.

See http://clang.llvm.org/compatibility.html#dep_lookup for more info.

BUG=none
TEST=sync_unit_tests

Review URL: http://codereview.chromium.org/3401002