Merge 60000 - Username+password weren't being saved for HTTP or FTP authentication.
HTTP + FTP authentication reuses the PasswordManager infrastructure for saving and autofilling passwords. When the user needs to be prompted for username + password, a PasswordForm is created. However, a username_element and password_element were never set in the PasswordForm.

The PasswordFormManager would then treat this form as invalid, and the values would never be saved even if the page loaded successfully.
 
The UI flow is a bit different from older versions of Chrome, but more consistent with other password saves: the user first presses the OK
button to provisionally save and try the new password. When the page loads successfully, an info bar than appears to ask the user whether the password should be saved.

BUG=5236
TEST=Go to site (or proxy) which requires Basic or Digest HTTP authentication, enter correct username+password and elect to save the password. Stop Chrome, reopen, and go to site (or through same proxy) and make sure username+password are autofilled in the form.

Review URL: http://codereview.chromium.org/3398010

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/3482002