Cleanup orphaned testserver processes on posix.
If a chrome test that uses a testserver were to crash, the testserver process ends up being orphaned, potentially affecting subsequent tests. This patch causes child processes of a test case to be killed along with the test case if it were to be abruptly terminated.

BUG=55808
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3423012