[Mac] Don't allow new I/O after profile dir is gone.
Mac tests have lots of failures in ImportantFileWriter.  It is because
the scoped temp dir associated with the testing profile is being
deleted before the background task on the FILE thread scheduled by
ImportantFileWriter is run.  Other places in the code simply don't
have a FILE thread, so the task is dropped.  This gets rid of those
threads before deleting the testing profile in ~BrowserTestHelper().

BUG=none
TEST=fewer nuisance log lines in unit_tests.

Review URL: http://codereview.chromium.org/3440012