Simple fix correcting the assignment of page_id to RenderView instances.
The logic in NavigationController::ClassifyNavigation assumes that all new pages will have an id LARGER than the maximal id that is locally cached in the SiteInstance of the active RenderViewHost.  

Because page ids are doled out by incrementing a static variable that resides in the render process - see RenderView::next_page_id_ -  resetting this counter while the render process is still in use will confuse the browser, and trigger the assert logic in NavigationController::ClassifyNavigation.  Newly created RenderView objects in this render process will send page ids  less than the value stored in SiteInstance::max_page_id during navigation operations.

This problem was manifested during the interaction of multiple instances of a chrome-extension residing in ActiveX controls.  Consider navigating a Chrome-Frame instance to an extension page.  Now navigate a second Chrome-Frame instance to another extension page within the same extension.  If a top-level link navigation is initiated on the second CF instance, then the extension will be torn down, but the old logic would have reset the next_page_id_ counter.  If a subsequent CF was navigated to a page within the same extension, it would be backed by the same extension process, yet it would be assigned an incorrect page id, as described above.

BUG=55138
TEST=none

Review URL: http://codereview.chromium.org/3442007