Refactor, rename and relocate ProfileSyncServiceTestHarness.
ProfileSyncServiceTestHarness needs to support a specialization that can use a WaitableEvent instead of starting and stopping the UI Message loop. Required for automation provider support for sync.

The class also needs to live in chrome/browser, instead of chrome/test, and could do with a rename.

BUG=56460
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3492005