Mark common instantions of CancelableRequestConsumerTSimple as extern templates.
This means that these template instantiations will be built once in cancelable_request.cc instead of in each of the 20 files that use it. The definitions have been moved to make them not inline (as explicit template instantiation only works with non-inline members).

This cuts more than 4 megabytes off libbrowser.a in Debug mode.

This patch does nothing when compiled with MSVC, due to compiler problems.

BUG=none
TEST=compiles smaller

Review URL: http://codereview.chromium.org/3520001