Disk cache: Change the way we delete files.
Now we always share delete and request delete access when we
open the files, and don't close the handle before deleting a
file. The result is that the delete operation should not be
prevented by someone opening the file right after we close
our handle.

The downside is that anybody attempting to open one of our
files while the browser is running will have to also share
delete or their request will fail. On the other hand, we
don't really want other people modifying the files at the
same time, so the change may be for good.

The caveat is that I think most of the failures are result
of improper AV behavior, so we have to see what happens.

BUG=16723
TEST=net_unittests


Review URL: http://codereview.chromium.org/3475025