Fix form_structure_unittest two-phase lookup problems.
The operator<< for webkit_glue::FormData needs to be declared
in the same namespace as its parameter type, in this case webkit_glue.

Otherwise Clang fails to compile this because it is more strict on the Standard.

See http://clang.llvm.org/compatibility.html#dep_lookup for more info.

BUG=none
TEST=unit_tests

Review URL: http://codereview.chromium.org/3583002