[dom ui content settings] show geolocation exceptions in the table
This works with the following caveat: the view does not update as the model does. The view does update the model (when pressing remove), but not vice versa (the view is only updated once, when it is first opened). This is actually also true of the native dialog version. To fix this I will need to add an observer interface to the Geolocation settings map similar to the one in HostContentSettingsMap.

I didn't use the GeolocationExceptionsTableModel because it doesn't seem helpful in this case, although I did copy some code from it. If we move to dom ui completely on all platforms, we can delete that file completely and the relevant parts will have migrated into ContentSettingsHandler. The GeolocationExceptionsTableModel allows for sorting, but I don't know if we care to keep that functionality as the mocks don't cover it, and if so it should probably be pageside.

content_settings_handler is getting more unwieldy and will continue to do so, which motivated me to refactor a little and in particular to add better (or rather, any) comments to the header. I may eventually want to factor out even more functionality as new, special content types get added (plugins, notifications, geolocation, etc? are all non-standard).

BUG=57457
TEST=manual

Review URL: http://codereview.chromium.org/3520012