GTK: Start implementing the new clear data dialog.
We don't block with a throbber while doing a clear data operation. The views
version of this dialog disables the entire dialog and displays a throbber while
waiting for an operation to complete. We don't do that yet.

This is important because there's no way to display error codes the sync server
might throw; it's comparatively likely that deleting server side data may fail
compared to deleting local data. That's the next step.

BUG=57877
TEST=New dialog has same functionality as before; new functionality intermittently works (since we can't display errors yet).

Review URL: http://codereview.chromium.org/3574013