Add a glyph drawing routine to private Pepper interface. The current implementation is Linux only.
This also changes the Private font interface to use the "regular"
PP_FontDescription_Dev struct for specifying the font, which allows us to unify
some of the data here. I kept the character set stuff as a separate parameter
which is needed for compat in a few cases.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/3569013