Fix flaky IncognitoSplitMode. For real this time.
I accidentally checked in the wrong change last time.

This reverts r62432 and commits the code that was actually reviewed in http://codereview.chromium.org/3656005/show .

Stop processing incoming messages in BrowserRenderProcessHost when it is about
to be deleted. See comment 20 on the bug for details.

BUG=53991
TEST=handled by existing test

Review URL: http://codereview.chromium.org/3772003