Make Run methods non-virtual on RRResolverWorker.
The goal is to turn on -Wnon-virtual-dtor warning for gcc. In this case, the virtual Run()'s do not appear necessary,
especially since they are private. If intended to be derived from, there is a "delete this" in the base class which would
point to requiring a virtual destructor.

BUG=45135
TEST=Target builds, validated that there are no derived classes of RRResolverWorker.

Review URL: http://codereview.chromium.org/3824002