Merge 62679 [2nd try] - This adds a plugin selection policy for selecting which plugin is allowed for a particular domain.
It is only used on ChromeOS.  It reads from a file that is installed
in a known location on ChromeOS, and uses that as it's policy.

When there are multiple plugins supporting the same mime-type, the
appropriate plugin file to load is now selected based on policy.

BUG=http://crosbug.com/7403
TEST=ran new unit test, tested on device.

Review URL: http://codereview.chromium.org/3717005

TBR=gspencer@chromium.org
Review URL: http://codereview.chromium.org/3814008