Move the rest of the Extension data into either StaticData or RuntimeData subclasses.
The ultimate goal is to move StaticData back up to the top level, so that
Extension itself will be a thread-safe immutable object. I plan on making Extension refcounted, and const everywhere that it doesn't need to be initialized. RuntimeData will be the only thing mutable, and then only on the UI thread.

BUG=56550
TEST=extensions still work

Review URL: http://codereview.chromium.org/3757001