Use TabContents for Constrained HTML dialogs.
This makes constrained html windows match HtmlDialogUI more closely. The main reason for wanting to do this is that TabContents{View} has a lot of functionality that we need, but miss out on if we use RenderViewHost directly. This isn't hooked up to anything yet, but can be tested using sync setup.

BUG=58022
TEST=manual

Review URL: http://codereview.chromium.org/3846002