Full refactor / rewrite of bookmark sync integration tests
The LiveBookmarksSyncTest and BookmarkModelVerifier classes were designed over a year ago and are badly in need of a refactor. Bookmark sync integration test cases are extremely cumbersome to write, review and read, with individual tests requiring a lot of repeated code.

This patch contains the following:
- A refactor of the BookmarkModelVerifier class.
- A major update to the LiveBookmarksSyncTest class, providing test cases with a bunch of new easy-to-use methods to perform various bookmark operations.
- A complete rewrite of every single bookmark sync integration test case using the new methods. (phew!)

In addition, this patch does the following:
- Replaces all instances of EXPECT_* in test cases with ASSERT_*, thereby avoiding several pages of irrelevant logs in case of failures, and making logs a lot easier to read.
- Annotates a few methods with WARN_UNUSED_RESULT so that test cases are forced to verify their results before proceeding.

BUG=59301,56043,19769
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3971003