Merge 63771 - [CHrome OS] HOSTED accounts will not be allowed to log in
Instead of allowing HOSTED accounts to log in, but be flagged, we are instead detecting them and blocking them.
Now, when a HOSTED account tries to log in, AuthenticateToLogin will call OnLoginFailure with USER_NOT_SIGNED_UP, indicating that the account in question is not allowed to use this device.

BUG=chromium-os:7867
TEST=unit tests, install on device and attempt to log in with a HOSTED account.  See that it fails.

Review URL: http://codereview.chromium.org/3973010

TBR=nkostylev@chromium.org
Review URL: http://codereview.chromium.org/4094003