Re-enable ExtensionApiTest.WindowOpen.
There was a bug in the original implementation of this test
that caused it to close early. Depending on the exact
timing, this sometimes failed, which caused it to be marked
flaky. Later, a bug was introduced that broke the underlying
functionality, which is also fixed here.

BUG=60156
TEST=Covered by browser_tests.

Review URL: http://codereview.chromium.org/4026005