Add logo display for fourth search engine in search engine selection dialog.
BUG=57993
TEST=Set the search engine on your (non-Chrome) default browser to be NOT in the top three for your region.
For USA, for example, you could use Baidu or Aol Search. Start Chrome with --first-run switch on an empty profile.
You should see the fourth search engine selection represented by a logo, not just text (if it's included in the list of logos for which we have images).
Review URL: http://codereview.chromium.org/4109001