This moves log output for ChromeOS to safer locations.
For BWSI (incognito) mode, we want to make sure that we don't leak
private information between sessions, so we need to have chrome write
log output to the tmpfs created for that mode.

Also, the chrome log in "logged in" mode should reside in the "log"
subdir instead of in the user data dir.

BUG=http://crosbug.com/6908
TEST=Ran on device in incognito and regular modes, and watched where
log output ended up.

Review URL: http://codereview.chromium.org/4194005