Change disabling the assert that is causing ExtensionApiTest.Popup to become flaky.
I've tried to reproduce the flaky behaviour on my local machine, in a VM, and on the build-bots, but have not been able to reproduce the problem.  Because this test is important, I am disabling the single assert so that the flaky label can be removed.

BUG=46601
TEST=ExtensionApiTest.Popup

Review URL: http://codereview.chromium.org/4264005