Added GPUBrowserTest.BrowserTestLaunchedWithOSMesa.
This just verifies that the OpenGL backend is OSMesa and does a simple smoke test that it is functional.

It is the browser test equivalent of GPUUITest.UITestLaunchedWithOSMesa. OSMesa should be required for both kinds of tests.

I moved both tests into the new chrome/test/gpu directory.

There are ongoing stability issues with OSMesa (and GL in general) on bots.

TEST=passes locally, try
BUG=none

Review URL: http://codereview.chromium.org/4424001