Detect corruption of previous snapshots in histograms
Having verified that histograms can be corrupted by
random memory smashers (or DRAM problems), this CL
looks at one last example of histograms that are
at rest for an extended period of time, and hence
vulnerable.  Between each UMA upload, we save
snapshots of the data we've already sent, so that
we can just send "new samples."  If those snapshots
are corrupted, the noise would be directly
injected into the UMA uploads.  This CL checks
for consistency in those snapshots, and if there
is any inconsistency, it skips over the
recent changes (since it has no baseline).

Since the code to do this was getting larger,
I factored it out a bit.  The hassle is that
I wanted to keep separate records of corruption
in the renderer vs browser, which complicates
the factoring a bit.

BUG=61281
r=mbelshe
Review URL: http://codereview.chromium.org/4733002