NSS: always record the ServerHello data.
Previously, if we predicted a ServerHello (because we were snap
startting), then serverHelloPredictionData was set and we wouldn't
record the actual ServerHello contents in the event of a misprediction.

This meant that we couldn't recover from a misprediction because we
would record the same prediction again.

With this patch, we'll always write to serverHelloPredictionData, even
if we made a prediction.

BUG=none
TEST=snap_start_unittests