Fix omnibox update bug.
When tab-switching happens, the state of the tab is stored so that it can be restored later. However, when a background tab is being closed, it doesn't cause a tab-switch, and trying to save the state causes problems in some cases (e.g. 61588). So do not save state of a background tab.

BUG=61588
TEST=see bug

Review URL: http://codereview.chromium.org/5075001