Fix a problem that would only show up in non-admin mode.
We were letting the result of IEIsProtectedModeURL decide the code flow of our tests, and for some reason, when run in non-admin mode, it causes a CoCreateInstance to be called and we mock it.

So we decided to mock IEIsProtectedModeURL too, so that we can have a more deterministic code flow, and now we have the opportunity (behind a TODO for now) to valid all code flows based on diferent return values.

TEST=Make sure the unit tests succeed in both and non-admin mode, which ever version of IE instance on whatever version of the OS.
BUG= 62387
Review URL: http://codereview.chromium.org/5116004