Cleanup histogram synchronizer
Some of the assertions in histogram synchronizer, specifying
thread-to-run-on, were breaking. We used to use a
fancy mix of threads and locks to access the data
involving which sequence was pending, and this
transitions to clean use of locking only for
all members of the class.

BUG=62810
r=wtc
Review URL: http://codereview.chromium.org/5189002