[GTK] prevent MenuBarHelper from referencing stale widget.
Menus can be destroyed without being hidden, so connect to the "destroy" signal as well as "hide". This fixes the crash, and makes the class more robust in the future.

Also, manually hide the bookmark menu widget when destroyed it in BookmarkMenuController. This gets the bbar button paint state correct.

BUG=63294
TEST=manual (see bug)

Review URL: http://codereview.chromium.org/5298002