Use BRANDING settings for IE CEEE. Simplify chrome_dll_version handling.
Adding new CEEE_PRODUCT_FULLNAME key to the BRANDING file. Adapting version.h.in to define a preprocessor constant for it. Using this and the other version/product name constants in IE CEEE's .rc files.

Piggybacking on MAD's recent changes to make %NAME% a variable in the .rgs files to fairly easily get the product name from branding in there. Naming the toolband with the product name directly (as it is what shows up in IE's View/Toolbars menu), the BHO with an added "Helper" and the broker with an added "Broker".

Removing the 'chrome_dll_version' target from chrome.gyp, and skipping code generation of chrome_dll_version.rc. Instead we have a source file with that name, that includes the verson.h generated by the 'chrome_version_header' target. This is cleaner and less code generation.

Fixing 'chrome_dll_nacl64' target, it seemed like it wanted to have version information, but it didn't.

Cleaning up the way sources are specified in several of the chrome_tests.gypi targets, to use fewer source! statements just to remove source files that are Windows-only (put them in the Windows-only section instead).

BUG=none
TEST=Visible change should be that user-visible names in Chromium builds are like "Chromium Extensions Execution Environment" (with "Broker" or "Helper" appended when appropriate), i.e. they lose the "Google" part.

Review URL: http://codereview.chromium.org/5360002