New Cloud Print Proxy setup UI.
New UI to setup the proxy in Chrome.  This UI passed review.  New features
include a friendlier layout, a Learn More link, and a Print Test Page button on
the success page.  This does not yet bring the UI out from behind a flag; that
will be a separate CL.

BUG=60092
TEST=CloudPrintURLTest.CheckDefaultURLs

Review URL: http://codereview.chromium.org/5285001