Cleanup X509CertificateMac style nits and introduce a crypto library agnostic means of parsing certificate dates.
In the process of addressing style issues, a bug was found in the date/time parsing routines used on Mac. Though not currently hit, it could lead to wild misrepresentation of the validity period for a certificate. Switch to using the routine used when using OpenSSL, which does things correctly.

BUG=none
TEST=existing

Review URL: http://codereview.chromium.org/4653002