Merge 67172 - Fix shutdown crash in CertVerifier by using a MessageLoopProxy.
The CertVerifier is not getting Cancel()'d because something which owns it is getting leaked, most likely a URLRequestJob.  Therefore, we can end up accessing a deleted MessageLoop on shutdown.  MessageLoopProxy prevents accessing a deleted MessageLoop on shutdown, instead it just deletes the task, which isn't great, but it's better than crashing.  We should fix the root cause eventually, which is a leak of the URLRequestJob.

BUG=42275,chromium-os:8179
TEST=none

Review URL: http://codereview.chromium.org/5347001

TBR=willchan@chromium.org
Review URL: http://codereview.chromium.org/5542008