Merge 68607 - Don't get tab id until tab window handle is ready.
There are rare cases where the executor manager will DCHECK that the window handle that we ask it to associate to our tab id doesn't have a valid IEFrame parent. This is because we send the request for id <-> handle association before the handle is completely ready... So wait until the tab window handle is valid before sending a mapping event.

BUG=65896
TEST=None

Review URL: http://codereview.chromium.org/5632004

TBR=mad@google.com
Review URL: http://codereview.chromium.org/5588010