[Mac] Work around FSEventStreamInvalidate() bug.
The way FilePathWatcherImpl on OSX uses FSEventStream caused a
double-close of a file descriptor.  If another thread allocated a file
descriptor during that window, it would later have problems.  This
mostly happened on browser_tests, but it could also happen in the real
world.

The change uses a more specific function to remove the stream from the
runloop.

BUG=66938
TEST=waterfall

Review URL: http://codereview.chromium.org/6048002