Set scoped IO access when closing the audio input thread to allow joining.
This is a fix for the crash problem described by issue 67806. The problem was caused by a new included IO access assertion that prevented the IO thread to join any other thread. This raised an error when trying to stop the audio recording thread.

BUG=67806
TEST=none

Review URL: http://codereview.chromium.org/6063004