Add a module to correctly import depot_tools.
Ironically, the depot_tools reference was relative and this is breaking the
commit queue for the commit queue since the patch to depot_tools was hard coded.
It now tries to be 'intelligent' by looking in PATH and falling back to walking
up the directory tree.

TBR=dpranke
BUG=
TEST=

Review URL: http://codereview.chromium.org/5986008