views (linux): Fix keyboard shortcuts.
BUG=Pressing backspace in a textfield in a web-page triggers the 'navigate back'
action in the browser. This is because the accelerator gets triggered before the
renderer has a chance to get to the key. To fix this bug, rearrange the keypress
event handler for RenderWidgetHostViewViews. The views-way of keypress handling
is more similar on windows and linux with this change.

TEST=see bug

Review URL: http://codereview.chromium.org/6002007