Full refactor and rewrite of the autofill sync integration tests
The autofill sync integration tests were written more than a year ago,
and are pretty badly out of date. In addition, the autofill test
framework is crufty and adding new tests is not intuitive.

The autofill sync code has recently undergone a major refactor, and so
it is time to update the integration test code too. This patch contains
a major update of the LiveAutofillSyncTest class and a rewrite of most
of the autofill sync tests.

BUG=62913, 60368
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6163001