Rework how bookmark bar folder menus and submenus are layed out when scrolling, adding and removing bookmarks and during drag and drop operations. Make a subfolder menu appear as much on screen as possible.
Changed to the BookmarkBarFolderWindow.xib: Added view (visibleView_) between the window's content view and the rest of the content. Added scroll-up arrow and scroll-down arrow views. Added connections to controller of all views within the window.

BUG=54701
TEST=Test submenu appearance by creating a folder menu with lots of items with the bottom-most item being a sub-folder with several bookmark items of its own. Place the browser window such that the bookmark button for the sub-folder will appear near the bottom of the screen. Pop up the folder menu then pop up the sub-menu and verify that the sub-menu is shown on screen and does not require scrolling.

In all following cases of adding or removing bookmarks, either by using the contextual menus (cut/copy/paste/delete) or by dragging and dropping, verify that the top of the folder menu and the menu items (bookmark buttons) above the location of the addition/removal remain in the same position vertically on the screen; menu items below the affected bookmark should move upward (for a removal) or downward (for an addition) and the scroll-up arrow (the one at the bottom of the menu) should hide or show as appropriate.

Verify that dragging or pasting a wide menu item into a narrow folder menu causes the menu to widen. Verify that cutting or deleting a wide menu item from a folder menu will cause the menu to narrow.

Perform the listed tests for each of the folder menu arrangements.

Folder menu arrangements
------------------------
1. Folder with no items (one showing 'empty')
2. Folder with just a few items and neither the scroll-up or scroll-down arrow showing.
3. Folder with enough items such that adding one more should cause the scroll-up arrow to show.
4. Folder with enough items and the scroll-up arrow showing such that removing one item should cause the scroll-up arrow to hide.
5. Folder with enough items that the scroll-up arrow shows but after fully scrolling the menu fits completely on screen without any scrolling arrows showing.
6. Folder with enough items that the scroll-up arrow shows and can be scrolled up such that both scroll-up and scroll-down arrows can show with plenty of extra items.

Tests
-----
1. Cut a bookmark menu item.
2. Paste an item.
3. Paste another item.
4. Drag an item to a different place within the same folder menu.
5. Drag an item to a different folder menu (one each of those listed above).
6. Cut the first item.
7. Cut the last item.
8. Paste to the last item.
For scrollable arrangements #5 and #6:
9. Scroll up only a little (such that the scroll-down arrow does not appear) and cut an item.
10. Scroll up only a little (such that the scroll-down arrow does not appear) and paste an item.
For scrollable arrangement #6:
11. Scroll up so that both the scroll-up and scroll-down arrows appear and cut an item.
12. Scroll up so that both the scroll-up and scroll-down arrows appear and paste an item.

Review URL: http://codereview.chromium.org/5694001