Disable background network access in sync integration tests
Now that the sync integration tests are designed to be self-contained
and not rely on external network resources like GAIA, we need to ensure
that background network access for things like auto update is disabled.

This patch adds the --disable-background-networking command line switch
to the sync integration test framework.

BUG=69469, 53931
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6282001