Add presubmit checks for grit.  This runs the unittests.
Fix the grd_reader tests to pass if run from any directory (previously, it
had to be run from tools/grit).

Fix rc_unittest to pass on Linux (it was sensitive to line endings).

If the user doesn't have access to the internal repository, there will be
lots of failures :(  Maybe there's a way to detect this case and skip the
tests?

Review URL: http://codereview.chromium.org/6349026