Update GYP to r882 to pick up r881:
The new postbuild step in Chromium's
third_party/WebKit/Source/WebCore/WebCore.gyp/WebCore.gyp:webcore_svg
is running before libwebcore_svg.a exists in clobber debug builds.
Explicitly declaring a target's intended output (linkable or static
library) as an input to the postbuild shell script action forces the
postbuild to wait for the link step to complete. Because no outputs are
declared, the postbuild step will still run each time the target is
built, which is the desired behavior.

This was previously attempted in r74191, but had to be reverted in
r74198 because it caused Windows clobber build failures. The offending
change to GYP, r880, was backed out at r882.

BUG=none
TEST=Clean Chromium build on Mac and Windows
Review URL: http://codereview.chromium.org/6467001