Add support for using ICU with a data file instead of compiling the data in.
This might help the Mac bots as it means each unittest that depends on base
will be ~10meg smaller, helping link and also helping with the copy of
bits from the builder to the tester.

If Mac ever support a Universal binary (32bit and 64bit) it will also probably
want to use this to avoid carrying two copies of the ICU data.

Review URL: http://codereview.chromium.org/6482029