This stops the sad tab favicon from showing when a tab is killed when the --reload-killed-tabs flag is specified.
This is so that when a tab is killed, we don't
indicate that the tab is dead, we just reload when the user selects it
(unless it's already the foreground tab).

BUG=chromium-os:8505
TEST=Ran on device, killed a non-foreground tab and observed that the
favicon didn't change.  Killed a foreground tab and observed that it still animated and changed.

Review URL: http://codereview.chromium.org/6312155