Fix service process to not idle-timeout if Chromoting is enabled.
When Chromoting is enabled via Chrome, the service process is started
if necessary and notified that the Chromoting service has become enabled.
When the service process is later auto-started (or run manually) it will
initialize Chromoting, which detects that it is enabled and starts up, but
fails to inform the service process that it's enabled, so the process waits
a while and then exits.

BUG=73168
TEST=Enable Chromoting in Chrome.  Kill the existing service process and start a new one manually.  Ensure that the new one runs for at least two minutes without quitting.

Review URL: http://codereview.chromium.org/6532023