Removing AccessLog class. This is a partial revert of yzshen's r74872.
The AccessLog class was something we added to track down a crash,
which is no longer happening after that change. Reverting piece by
piece and waiting for the next canary build before doing the next
part.

Note that the use of the AccessLog class could have introduced slight
serialization to the method
URLRequestThrottlerManager::RegisterRequestUrl as it would call the
AccessLog::Add method which synchronized on a lock. So removing is not
100% safe; I will watch for crashes in the canary builds after this.

BUG=71721
TEST=none

Review URL: http://codereview.chromium.org/6538061