Only fail in extract build if asked to do so.
Check first if halt_on_missing_build is present in
extract build.  This fixes a case where either a
slave is still using the dependent scheduler (which
can sometimes end up with the wrong revision number)
or a scheduler that's watching another tree (which may
not be the revision used in creating the build zip).

Review URL: http://codereview.chromium.org/6543026