The RenderView currently only updates the navigation state for the currently displayed data source in OnDisplayPrerenderedPage.  If there's a provisional data source, it still thinks its being prerendered when it becomes the main data source.
This CL fixes that issue.  To do this, it moves the
|is_prerendering_| flag over to the RenderView.  While
not a strictly necessary design change, things are a bit
less bug-prone this way.

Also changes how |was_started_as_prerender| is set.

BUG=73173
TEST=none, yet
(manual test:  Go to http://peter.sh and refresh it a few times, with prerender enabled and a debug build.  Notice lack of DCHECK assertion in OnDisplayPrerenderedPage)
Review URL: http://codereview.chromium.org/6532031