Fix a build failure with bundled libvpx.
This is part of an effort to allow building with system libvpx.

I was only using trybots previously, so not all problems
got detected during the upstreaming, especially with
bundled libraries.

This is not ideal, but I just copied the include_dirs
snippet from elsewhere.

TBR=fischman, scherkus

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/6602087