Record histograms for Prerender control group.
The histograms were not being recorded since PrerenderPLTRecorder was checking for existence of a PrerenderManager in the current Profile, and one is not present for the control case.

I fixed this by making the PLT recording functions static: they already had no instance-specific dependencies, so this works pretty easily.

BUG=None
TEST=Force Chrome to be in prerender control group [requires modification to prerender_field_trial]. Do a query for "Google Trends" and click on it. about:histograms should have Prernder control page load times and windowed page load times.

Review URL: http://codereview.chromium.org/6597109