Make sure to remove RLZ-related registry values on chrome uninstall.
BUG=3486114
TEST=Install chrome with a non-organic brand code.  Start it and wait for
the RLZ registry values to be written to HKCU\software\google\common\rlz.
Uninstall chrome and make sure that the registry values are gone. Note
that only the registry keys and value that start with the letter 'C' relate
to chrome; any other that may exist after uninstall are OK.

Review URL: http://codereview.chromium.org/6597089