Mac/Linux proxy fix.
On Mac/Linux we keep the socket open, so don't get an OnChannelConnected
message when a new Chromium launches and connects to the service, so we weren't
properly marking ourselves as having a client connected. When the service was
disable and Chromium exited, it meant we weren't properly exiting
immediately. This is only a partial fix for the bug.

BUG=74983
TEST=Start Chromium, enable the proxy, quit Chromium, start Chromium, disable the proxy, quit Chromium. With this fix the service process immediately exits.

Review URL: http://codereview.chromium.org/6635023