Adding an image to avoid future try server failure.
This is for the upcoming redesign changelist for the
redesign of the Manage Extensions page (as a Settings
page).

Try servers tend to fail (at least have in the past)
if a binary file was included in the changelist
being 'tried'.

This should avoid that problem.

BUG=52447
TEST=None
TBR=mwichary

Review URL: http://codereview.chromium.org/6652021