Remove TODO in AppCacheURLRequestJob about GetContentEncodings().
I plan to make some changes to the interactions between URLRequestJobs and Filters,
and so as a convenience I'd like this file to not show up in code search/git grep.
It seems safe to assume that the URLRequestJob is already correctly implemented,
as this comment has been here for well over a year with no change.

R=michaeln@chromium.org
BUG=none
TEST=none


Review URL: http://codereview.chromium.org/6698050