Stop subclassing FilterContext in URLRequestJob.
Create a wrapper for URLRequestHttpJob that subclasses FilterContext,
and pass an instance of this facade to Filter::Factory().

Reduce the scope of URLRequestJob's interface as much as possible, moving
methods into URLRequestHttpJob or its facade as appropriate, and making
methods that remain in URLRequestJob non-virtual where possible.

BUG=none
TEST=net_unittests,try bots

Review URL: http://codereview.chromium.org/6677104