External extensions should not be installed if the ID or version in the external source is wrong.
Ignoring incorrect IDs makes it impossible to keep an uninstalled external extension from re-installing.  On uninstall, we record that an extension was uninstalled using its ID.  If the external source claims the extension has a different ID, it will not match that record, and chrome can not see that the extension should not be installed.

BUG=75355
TEST=ExtensionServiceTest.FailOn*

Review URL: http://codereview.chromium.org/6713030