Revert 79352 - ChromeFrame would fail to load the page correctly if a manifest was specified in the html tag. This would work the first time and fail on subsequent attempts. Basically the manifest specifies portions of the website that is cached locally. In this case when there is a subsequent navigation to the original URL the cached resource can be served out of the appcache which lives in chrome.
ChromeFrame expects the first url request from chrome to match that of the url which the user
navigated to. This url is served out of chrome'a app cache thus causing the pending url reques
checks to fail.

Fix is to validate the pending url request with the request received and stop the pending request
if there is no match. We just trace the fact and stop this request. A new request object is 
created to handle the request received and life goes on.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=76439

BUG=76439
TEST=manually as described in the bug.
Review URL: http://codereview.chromium.org/6713112

TBR=ananta@chromium.org
Review URL: http://codereview.chromium.org/6741001