Be more thorough checking for NULL NPP values
Unlike the other entry points, NPN_GetValue and NPN_SetValue don't check
for NULL npp values. nspluginwrapper will happily call functions with
NULL npp if the instances have been destroyed in the meantime. (And the
patch for #53940 will make this happen more often to avoid a plugin-side
crash.)

NOTE: This does not fix #53940, but the fix for it in nspluginwrapper reveals
some missing checks on our end.

BUG=53940
TEST=none

Review URL: http://codereview.chromium.org/6722021