Fix resource destruction in proxy
This ensures that the resource on the plugin side is destroyed before we send
the message to the host, so that it has a chance to do proper cleanup.

Also, fix Surface3D destruction that could cause a write-after-free.

BUG=none
TEST=go to youtube with out-of-process pepper flash. click on fullscreen.
observe no hang, no crash

Review URL: http://codereview.chromium.org/6771042