net: add support for checking if a known public key is in a chain.
This is intended for use with future support for HSTS or CAA pinning of
certificate chains. Since we build certificate chains from a pool of
certificates, and certificates get reissued with updated expiries etc, only the
public keys in the chain will always match. Thus we match on SHA1 hashes of
SubjectPublicKeyInfos.

BUG=none
TEST=net_unittests

Review URL: http://codereview.chromium.org/6821010