Support decoding GenericNamedValue based policy.
Even if it arrives in the ChromeSettingsProto message where Chrome expects to find explicitly typed policy.
This is necessary for as long as CPanel/D3 deliver old-style policy. We include this fix on the client side to make a corresponding workaround in DMServer unnecessary.

BUG=chromium-os:14102
TEST=UserPolicyCacheTest.OldStylePolicy; manual test against DMServer without the server-side workaround

Review URL: http://codereview.chromium.org/6840014