New automation hook to set the app launch type (e.g., window vs. tab).
This hook allows the launch type for a specified app to be set as either
a regular tab, a pinned tab, a fullscreen tab, or a tab in a new window.
Sample PyAuto tests are included that exercise this hook.  The GetBrowserInfo
automation hook is also slightly modified to return information about
whether each existing tab is pinned or not.

BUG=75205
TEST=None.

Review URL: http://codereview.chromium.org/6852005