Factor a scheduler object out of MetricsService.
The logic for when to upload is now separate from the core of MetricsService. For now the object is internally constructed so that the MetricsService interface doesn't change, but it could easily be pulled out later if we need DI-style construction.

Fixes a bug where backoff for server errors would be bypassed if the logs were large.

Also removes a bit of dead code I missed when removing the server response parsing in the last CL.

BUG=None
TEST=Metrics continue to show up on the dashboards unchanged.

Review URL: http://codereview.chromium.org/6869034