Use large icon resource pak
This is a part of r82185 that was reverted. The change was reverted because it caused a performance regression.

In r82538 and r82584 I re-checked in code to split high-res icons into a seperate resource pak and load it.

This change adds code to actually use use the high-res icons.

BUG=None
TEST=None
Review URL: http://codereview.chromium.org/6897013