Stop reference counting BaseSessionService.
BaseSessionService already uses ScopedRunnableMethodFactory to post its
tasks, so there is no need for it to be reference counted as only
Profile will ever have a reference on it. It holds a pointer to the
Profile so it's unsafe for it to continue to exist after the Profile's
destruction in any case.

BUG=79886
TEST=existing tests

Review URL: http://codereview.chromium.org/6877043