Use large icon resource pak
This is part of change r82185 that was reverted due to a performance regression on 10.5.

The performance regression was due to a change in skia_utils_mac.mm to draw images using -[NSImageRep drawRect:] instead of -[NSImage drawRect:...].

It turns out that on 10.5 -[NSImage drawRect:...] does caching that imporoves performance for subsequent drawing by as much as 4x.

This change is similar to r82185 expect that we now use -[NSImage drawRect:] if available.

BUG=75812
TEST=Ran unit tests locally, try jobs pending.
Review URL: http://codereview.chromium.org/6905003