Initial inclusion of Proxy tests
Initial inclusion of Proxy tests

The following tests will test that the Proxy settings are saved properly
using different types of input.  It does not test that the fields are actually
functional and that the device is using the proxy.  That will come later.

BUG=None
TEST=Run it

Review URL: http://codereview.chromium.org/6914013