[Sync] Support for non-blocking conflicts and encryption conflicts.
This patch introduces the notion of non-blocking conflicts. These are conflicts that do not result in the
syncer getting stuck. They affect the HasConflictingUpdates status call, so we attempt to reapply updates
when they occur, but they are not passed to the conflict resolver, and hence do not result in the syncer being stuck. All cases where we have a conflict due to encryption/decryption result in a new 
ENCRYPTION_CONFLICT, which are treated as nonblocking conflicts.

Updated sync's protocol_version to 27 so the server can tell which clients understand encryption.

Also, due to this change, it's more likely that local/server changes will be overwritten in the conflict
resolver (since encryption conflicts are kept as unapplied while they can't be decrypted). Added
counters for how often this happens, which will appear on about:sync, and filed follow up bug
crbug.com/76596.

BUG=59242
TEST=new session sync_integration tests, modified apply_updates_command unit tests


Review URL: http://codereview.chromium.org/6714002