Set minimum width on menu for NativeComboboxViews.
The combobox menu is forced to be at least as wide as the combobox itself.

Besides the addition to native_combobox_views.cc, an edit was needed to views::SubmenuView.  The minimum preferred width must account for the border (specified in SubmenuView, applied in MenuScrollViewContainer) to make the ultimate menu width match the size requested by the combobox.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7002004