Partially fix WGL backend on windows.
It should render on all windows boxes with working GL. The GPU process creates a child window to render to again, but only if desktop GL has been enabled. The child window is not created by default so there should not be issues with deadlocks. Resizing flickers for some reason.

TEST=WebGL works, trybots.
BUG=none
Review URL: http://codereview.chromium.org/6991002