Drive compositor scheduling with the swapbuffers callback.
This modifies the compositor to hold DoDeferredUpdate until the previous frames are ack'd by the GPU.

This vastly improves rendering smoothness.

There is some complexity in the way we intercept the SwapBuffers callback due to the different types of GraphicsContexts that get scheduled by render_widget.

BUG=80480,53051

Review URL: http://codereview.chromium.org/6902106