Do full-screen polling if X server does not support DAMAGE.
Not optimized yet - need to detect when the screen image is unchanged and avoid
passing it to the encoder (as in the Windows implementation).

BUG=None
TEST=Manual

To test:

Switch VT (Ctrl-Alt-F2) and run

startx -- :1 -extension DAMAGE

Run chromoting_simple_host from there, and connect to it.  Verify that changing
screen resolution works.  Also stress the X Server with

x11perf -time 99999 -osrect 500

and repeat the screen-resolution test.

Review URL: http://codereview.chromium.org/7012025