Re-enable integration tests that completely disable/enable sync
The integration tests that completely disabled / enabled sync were
occasionally failing on the chrome buildbots, but not on the trybots or
local workstations.

r85503 fixed the underlying issue by restructuring the methods that
enabled / disabled sync on a machine.

This patch re-enables the tests that were earlier marked as FAILS_.

BUG=82489
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/7030021