Make clicks above status area activate status area buttons.
BrowserFrameViewChromeos adds a few pixels of pad to the top of the
status area. For mouse events in this area to activate a button,
hit testing need to be successful. We do this by clamping y to the
position of the menu button view in StatusAreaButton::HitTest.

BUG=chromiumos:5416
TEST=none

Review URL: http://codereview.chromium.org/7030005