Try another way to alias a variable in optimized builds.
The previous way did not fool LTCG optimization.

I tested that this works by doing and LTCG build without this change and verified that the compiler strips out the assignment to program_counter in MessageLoop::RunTask. Then I repeated with this change and verified that the compiler did not strip it out.

TEST=compiles plus the above
BUG=81449
Review URL: http://codereview.chromium.org/7054025