Register with Keystone using a versionPath and versionKey.
Providing this data to Keystone will cause it to check the application's
Info.plist to determine the installed version, rather than using the version
embedded within the ticket. The version in the ticket can become stale in
some rare situations through innocent user action.

This allows a user to replace Chrome with an earlier version (perhaps on
another channel) and not get "stuck" without autoupdates until that channel
catches up with the version listed in the ticket. (b/2506062)

This also prevents Keystone from believing an update is still available if
Chrome is updated but not restarted and the user subsequently promotes
Chrome's user ticket to a system ticket from within Chrome. (crbug.com/38696)

BUG=38696
TEST=On a user ticket, ksadmin -p should show a versionPath and versionKey
     after Chrome is launched. If promoted to a system ticket, ksadmin -Sp
     should show these same two parameters.
Review URL: http://codereview.chromium.org/6990046