Fix pyauto launcher for chromeos
PyAuto login automation is broken such that it crashes the browser in cases where the ownership is not taken and you don't wait long enough for chrome to startup, login dialog to show, etc. It takes time for chrome to sink these in.

For this reason pyauto tests on autotest lab machines were failing (the autotest lab machines begin with a clean stateful partition, and then we trigger a session_manager restart after creating enable_chromium_minidumps file)

This CL fixes the launcher.

The Login() automation still needs to be fixed to be resilient to the login dialog to show up.

BUG=15582

TEST=pyauto test fire up on chromeos autotest lab machines

Review URL: http://codereview.chromium.org/7065016