Add initial framework for performance tests that measure frame rate.
Add the "blank page" test, which just scrolls a page with a very large body.
The actual frame rate measurement is based on the webkitRequestAnimationFrame
callback.  The framework records how frequently that API generates a callback.
If it generates a callback at 60 FPS, then we know nothing is slowing us down.
This of course assumes that webkitRequestAnimationFrame is well implemented.

The advantage of this test framework over one that directly reads a framerate
from Chromium is that this one works outside of the test harness.  You can just
navigate your browser to the test file and run the test manually.

R=brettw,nduca
Review URL: http://codereview.chromium.org/7038052