pyautolib: shared_library -> loadable_module
shared_library is for libraries used within Chrome.
loadable_module is for general .so files we intend to be loadable.

loadable_module ends up in the product dir on Linux anyway, so
we can remove the symlink step.  .so is the extension for
loadable_modules on Mac, so we can remove the explicit name
change.

Review URL: http://codereview.chromium.org/7062027