Read int properties from the X server as long.
The documentation says that the values are returned as long,
even though the range of values is only 32-bit.  This makes
it safe to cast from 64-bit to 32-bit.

GetIntProperty isn't used anywhere and GetIntArrayProperty is
only used on ChromiumOS, so it never hit this bug before.

Review URL: http://codereview.chromium.org/7020010