Let MSys PATH conversion work.
Don't specify the full path to the EDITOR since a
DOS path will fail to work under MSysGit + SH + Git.
Instead, specify a relative path (just the executable
name) and add the absolute path to the PATH.  We then
rely on MSysGit to convert PATH to a format the
executables will understand.

Add MANIFEST.chromium to the repository and copy it
when creating the git_bin.zip archive.

Increment the VERSION specified in MANIFEST.chromium
to 2.

This change will require respinning git_bin.zip.
Review URL: http://codereview.chromium.org/7107002