Prevent a NULL deref in NTP tab restore.
If reopenTab is called on the new tab page before the list of closed
tabs has been retrieved, the tab restore service is still NULL and gets
dereferenced. This can only happen if the user injects extra JS calls
using devtools or similar, but it's easy to check for so we will check
for it to prevent the crash.

BUG=77877
TEST=


Review URL: http://codereview.chromium.org/7020023