make command buffer work in DRT
Note several tests get different results than the non-command buffer path.
I'm not sure if that's bugs in this impl, bugs in the command buffer
or bugs else where but that's kind of the point. This CL is supposed
to help find those bugs.

Is there something I should do to mark them to be fixed before
checking this in? Should I leave webkit_support.cc out of this
CL?

TEST=ran layout tests
BUG=84157

R=jamesr@chromium.org


Review URL: http://codereview.chromium.org/7085002