Allow delete of never-opened -wal file in Incognito.
The recent SQLite import added a feature called Write-Ahead Logging.
It has to check and respect a -wal file when opening the database.  If
the database is empty, the -wal is always deleted, without being
opened first.  This allows deletes of never-opened files in Incognito
mode to succeed.

BUG=84011
TEST=See bug, especially comment #10.

Review URL: http://codereview.chromium.org/7056035