Clean up base/Version
It turns out base/Version is really a value object but probably because of some serious accident or by
the machinations of a super villain, forgot his identity and now it thinks is a reference object,
only creatable in the heap and that could only spawn offsprings via cloning.

But fear not 'cause I've seen Version true nature; At its core is just a good 'ol vector<uint16>, which has
very respectable value semantics. Also I have removed the is_valid_ parasite as much as I could.

The old interface (GetVersionFromString and Clone) is kept so existing callers would not need to be modified.

BUG=none
TEST=included
Review URL: http://codereview.chromium.org/7105008