views: Use vector<T>::operator[] instead of vector<T>::at in the remaining cases.
The implementation of vector<T>::at performs bounds checks and can through an
exception std::out_of_range if you try to access an element that is not present
in the array. That is less eficient and we don't want to through an exception
anyway.

BUG=None
TEST=None

R=pkasting@chromium.org,sky@chromium.org

Review URL: http://codereview.chromium.org/7125012