file manager: clean up rename click detection
This patch changes rename click detection so that a user must slow double click on the *same* file.  The previous code usually required users to click on the same file, but relied on a side effect of the list control (that clicking on a new file resets the selection), which didn't always work.

BUG=chromium-os:16270
TEST=manual testing on chromeos-chrome on linux

Review URL: http://codereview.chromium.org/7134004