Avoid a NOT_REACHED() hit when we get an empty 404 page and link doctor is down or disabled.
We weren't fully specifying WebURLError, which caused us to
be confused when looking up the strings for the error code.

We still don't have a nice error message for this, but at least
we don't DCHECK in debug and this will now at least cause the
local error page to say there was an http 404 error.

BUG=76457

Review URL: http://codereview.chromium.org/6995091