Initial inclusion of ForgetWifiNetworks
When a user successfully connects to a network, that network is added to a remembered_wifi_networks list which holds a wifi object that allows users to autoconnect to the network in the future.

ForgetWifiNetwork serves the functionality of "Forget Network" button that appears for each network in the remembered networks page.

* Adding support for Forgetting Wifi Networks.
* Added remembered_networks to the GetNetworkInfo dictionary
* Fixed a bug with the service order comparison.  Should have used intersection and not union.

BUG=None
TEST=Run chromeos_wifi_compliance.py and chromeos_wifi.py

Review URL: http://codereview.chromium.org/7046104