net: don't check revocation when fetching PAC files.
If a PAC file is configured on an HTTPS URL we get into trouble. In order to
check revocation we need to make an HTTP request to the OCSP/CRL server, which
needs the PAC script to load and so we deadlock.

With this change we don't check revocation for PAC fetches.

BUG=86219
TEST=Configure a PAC script on HTTPS.


Review URL: http://codereview.chromium.org/7170026