Merge trunk r89741 to the 13.0.782 branch.
Get rid of ClearByPathBackupExclusion.

We planned on keeping ClearByPathBackupExclusion around as temporary code to
clean up the mess we used to make in
/Library/Preferences/com.apple.TimeMachine.plist. Evidently, when we ask
CSBackupSetItemExcluded to include-by-path an item that was previously
excluded by path, it will remove the ExcludeByPath entry for the path from
that plist, but it will also add an IncludeByPath entry. There is no way
using this API to clean up the mess we made in the past, we can only trade
an "exclude" mess for an "include" mess. It's not worth attempting to modify
the file ourselves, avoiding the CSBackupSetItemExcluded API, and it's
probably not even possible because the plist will likely be rewritten by
a system process anyway.

BUG=86845, 85999
TEST=No changes made to the plist mentioned above.
Review URL: http://codereview.chromium.org/7206030

TBR=mark@chromium.org
Review URL: http://codereview.chromium.org/7212016