Fix Pepper plugin sandbox interface for out-of-process plugin.
On Chrome OS Flash player is running out-of-process by default,
and it is failing to handle font fallback because in PpapiWebkitClientImpl
SandboxSupport object is not instantiated.

Without SandboxSupport instance, all the font related requests are handled
locally, and fail to communicate with Fontconfig because it is sandboxed.

Instantiating SandboxSupport will properly route font related requests
to Sandbox IPC process to communicate to Fontconfig.

BUG=87039, chromium-os:15882
TEST=Set Chrome OS to English UI and open http://weathernews.jp to see if
all the characters are rendered properly.

Review URL: http://codereview.chromium.org/7273062