ChromeFrame should report anchor navigations to Chrome. Currently this does not work as expected if a navigation is first attempted to a ChromeFrame URL without the anchor followed by a navigation to the same url with an anchor. In this case IE first sends an unload request to the currently loaded CF document which causes a lot of grief as we run unload handlers on the page. The navigation never proceeds to the new url though.
The proposed fix is to handle this in the onunload handler in CF by validating whether the new url
without the anchor is the same as the current url. If yes we just initiate a new navigation in Chrome.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=85617

BUG=85617
Review URL: http://codereview.chromium.org/7290015