Added ability to run a command, getting output and exit status on POSIX.
Register protocol handler needs to check the return value of scripts it runs and
collects output from. There needs a variant of GetAppOutput which returns the
exit code of the application run, not just a success or failure, so this function
has been added.

BUG=83557
TEST=Unit test added


Review URL: http://codereview.chromium.org/7198034