Merge 90410 - Add MenuItemView API to control menu repositioning.
The use case for this is the ChromiumOS OOBE network menu.  This menu is dynamic - access points may be added or removed while the menu is open so the number of items in the menu can change.  At a certain threshold, the menu will "hop" between above and below the button.  This CL provides a way to disable repositioning so the menu is always displayed below the button.

BUG=chromium-os:16720
TEST=open OOBE network menu with a lot of WiFi access points around, check that the menu is drawn below the button


Review URL: http://codereview.chromium.org/7217026

TBR=rhashimoto@chromium.org
Review URL: http://codereview.chromium.org/7316002