Remove the background page, and majorly refactor Javascript + Webapp.
This removes remoting_session.html, puts the "in-session" UI into the same page, uses classes to hide/display elements, and redefines what we mean by the <section> tag.

Also takes a bunch of the javascript logic and makes classes out of it as well as detangling a bit of the UI manipulation from the behavior functions.

BUG=none
TEST=manual


Review URL: http://codereview.chromium.org/7277020