Mac only: Change launchd type of service process.
There seems to be a (long-standing) issue where launchd agents marked for the
'Background' session type still get a terminate when the user logs out, but
don't get re-launched when the user logs in (the entry doesn't get unloaded
from launchctl on log-out, so the launch on load isn't effective). This works
around that problem by changing the session type to Aqua, so launchctl does
indeed unload the service process when the user logs out, and re-loads it when
the user logs back in again.

This is one workaround to the bug, the bug contains the other possible (more
heavyweight) one - I think this one is correct for the current service process
usages.

BUG=88954
TEST=Enable cloud print proxy on Mac, reboot, log in, check for service process, log out, in again, check for service process. Should still be running. Repeat log out/log in again to be sure.


Review URL: http://codereview.chromium.org/7336001