Make header copyable in views info bubble; add a comment about description.
This change makes the info bubble text copyable by changing it from a Label
to a Textfield (similar to how it's done in about_chrome_view.cc for the 
version text). Unfortunately, the same can't yet be done to the description
label since currently a textfield cannot be multiline. I've added a comment
to that extent.

BUG=52526
TEST=manual

Review URL: http://codereview.chromium.org/7316022