Add COMPONENT_BUILD global define.
This avoids the need to define FOO_DLL macros for each project that we wish to
optionally build as a DLL (when component=="shared_library").  This in turn
means that we do not need direct_dependent_settings to define FOO_DLL, and that
means that we don't need to update projects to convert transitive dependencies
into explicit dependencies.  This makes the component build more consistent
with the static build.

An alternative would be to use all_dependent_settings, but I feel that the
global approach is simpler as it creates less repetition in each target
definition for components.

A side-effect of this change is that I needed to make base_nacl_win64 be a
shared_library in the component build.

R=rvargas,bradnelson,evan
Review URL: http://codereview.chromium.org/7344022