Improved the WebGL conformance tests.
If the test now results in a timeout, it fails. Before, it would erroneously
pass. Also, the test first checks if it can create a WebGL context, failing
immediately if it cannot (what's the point of a WebGL conformance test
without WebGL?).

When a test fails, it now prints informative, machine-specific WebGL data.


TEST=gpu_tests --gtest_filter=WebGLConformanceTests.*


Review URL: http://codereview.chromium.org/7359004