[Sync] Fix session sync intetgration tests.
Refactor various parts of the sync integration tests. We now ensure that the
local session sees the open tab and wait until the syncer sees the changes
before continuing. In addition, the session change processor now listens to the
LOAD_COMPLETED_MAIN_FRAME notification, as before we were not catching when a
tab finished loading (which is when the title gets set).

BUG=85294
TEST=session sync integration tests succeed

Review URL: http://codereview.chromium.org/7277055