Initial sensor event provider for testing views screen rotation
The sensors API will unify various types of sensor data into a set of channels, each of which provides change events and periodic updates.

This version of the API is intended only to support the experimental screen rotation code and is not for general use. In particular, the final listener will declare generic |OnSensorChange| and |OnSensorUpdate| methods, rather than the special-purpose |OnScreenOrientationChange|.

BUG=none
TEST=manual with http://codereview.chromium.org/7273073/


Review URL: http://codereview.chromium.org/7366011