Missing changes for IDL
To minor changes to comments for the IDL.
Updated pp_stdint.idl so that we could generate it instead of checking in pp_stdint.h.
In addition it now provides a location for defining built in types.  Eventually we
can migrate type properties into this file.
Review URL: http://codereview.chromium.org/7448005