Add debugging messages to diagnose flakiness on Chrome OS.
FastShutdown.SlowTermination seems to be flaky (timing out) on Chrome OS. I've
been trying to reproduce the failures on the try bot (linux_chromeos) for an
entire week now, but without success. This might be because the try bot runs
debug, and not release. I don't have access to any other CrOS devices. So this
CL adds temporary debugging to the code; once the root cause of the timeouts
has been isolated, I'll remove the debug messages.

BUG=89173
TEST=none


Review URL: http://codereview.chromium.org/7465003