Delete TabContents::OnCrossSiteResponse
This code was never called, because TabContents doesn't implement
the interface that calls this method.  I assume it was moved to
the class this was forwarding to.  I tried to track down when
this happened in the history but we've moved code so many times
I couldn't find it.

Review URL: http://codereview.chromium.org/7467009