sync: listen for missing notifications
This adds a 'Could not connect' error if the server is unreachable instead of infinite spin.  Trying to sign in again won't work though, because there's no path to re-attempt backend initialization.  I'm going to add a way for SBH to callback on failure as well as success in an upcoming patch, but it's separable from this change.

BUG=87871,85022,88109
TEST=SyncSetupWizardTest, ProfileSyncServiceStartupTest

Review URL: http://codereview.chromium.org/7464005