Fix views_examples NativeWidgetViews parenting on Windows.
The problem this CL addresses is that the parent for a NativeWidgetViews-backed widget must be specified as a Widget and not a View.  This produces inconsistent results for views_examples because on Linux each tab is a Widget but on Windows each tab is only a View.  The result on Windows is that the NativeWidgetViews test goes on the 0th child of the Widget contents, which is the check box tab.


BUG=chromium:88716
TEST=run views_examples on Windows, NativeWidgetViews should be on proper tab


Review URL: http://codereview.chromium.org/7517015