[Mac] Fix BMB folder menu being closed on sliding animation end
The cause was a case where |hoverButton_| did not correspond to
the currently open folder, causing +draggingEntered: to close
the folder instead of opening it (or specifically, instead of doing
nothing in this case, since the folder to be open was the same).

The CL also has a second fix that prevents the menu from sometimes
being closed at certain mouse positions. This fix makes
-buttonForDroppingOnAtPoint: return the |hoverButton_| if the point
is within its bounds, rather relying on the "middle half" bounds
matching, which is done otherwise.

BUG=90846
TEST=manual - see bug

Review URL: http://codereview.chromium.org/7526003