[Sync] [Jank] Clear the vector of verified updates after processing.
Without this change, the list of verified updates would include all previous batches of downloaded updates.  So the Process step for the nth batch of updates would process all updates from previous batches.  This behavior was O(n^2).  We never noticed because the Process operation is idempotent.

I discovered the performance problem while torture-testing using a hacked server implementation that returned redundant updates 98% of the time, leading to very long getupdates cycles.  This seemed to cause noticeable UI jank as time went on; the UI thread was blocked for as long as 200ms.

TEST=hold down the 'a' key while typing in URL bar while initial sync download happens for an account with tens of thousands of updates.  Key should repeat smoothly without jank or hiccups.


Review URL: http://codereview.chromium.org/7519019