Move Var and StringVar to the ppapi/shared_impl so they can be used in the proxy. This does not, however, actually use these in the proxy yet, I'll do that in a subsequent pass.
ObjectVar is now renamed NPObjectVar and continues to be in webkit/plugins
since it uses NPObjects. There will be a corresponding future class in the
proxy for proxied objects.

The TryCatch that was in var.h is now in npapi_glue.h

Most of the rest of the change is updating namespaces.

TEST=unit tests
BUG=none
Review URL: http://codereview.chromium.org/7488062