Preload with RVH/RV messaging instead of NOTIFICATION_NAV_ENTRY_COMMITTED.
This patch is the first working version using a preload message in the style of
viewsource.

NOTIFICATION_NAV_ENTRY_COMMITTED was too late as it was after the navigate
command was sent and was a race condition to inject before the onload ran.

Basic gist of this patch is:

- Wait for RVH creation by listening for NOTIFICATION_RENDER_VIEW_HOST_CREATED.
  Since this is called from constructor, we must listen to AllSources() and
  assume the next RVH creation is ours. This is controlled test case, so we're
  probably ok.
- When RenderViewHostInitialized, call the OnJsInjectionReady passing the
  RVH. JsInjectionReadyObserver should now use that RVH to Send a
  ViewMsg_WebUIJavascript message to register javascript to preload.
- the ChromeRenderViewObserver will stash the javascript and load it when
  DidStartLoading.

BUG=90476
TEST=browser_tests --gtest_filter=PrintPreview*WebUI*.*

Review URL: http://codereview.chromium.org/7531029