Merge 95141 - Alwas set saved maximized state for panels to 'false'.
For reasonss that are not entirely clear, Browser::GetSavedMaximizedState() is sometimes setting maximized to true for panels created by an extension on ChromeOS.
Since panels have no concept of a maximized state, fix this by always returning 'false'.

This also includes a patch to widget.cc that is not necessary to fix this behavior, but it fixes the case where ShowInactive gets called when the widget is maximized (which shouldn't normally happen).

BUG=chromium-os:18534
TEST=See issue + test opening panels from an extension

Review URL: http://codereview.chromium.org/7551002

TBR=stevenjb@google.com
Review URL: http://codereview.chromium.org/7508045