Update perf dashboard overview.html
The linux graphs on the perf dashboard overview were pointing to an out of date url, and as a result were displaying out of date information.

This change fixes the linux graph urls, and also also makes the sync graphs correctly match up with the bots that actually run the sync perf tests.
Review URL: http://codereview.chromium.org/7574022