Revert 95441
This attempted fix for the bug referenced below did not help.

Original message:

TaskClosureAdapter::Run does not invoke the task if it is null.

This is an attempt to fix http://crbug.com/81449. If this actually fixes it then there is probably something more fundamentally wrong.

BUG=81449

Review URL: http://codereview.chromium.org/7598015