Remove some unneeded uses of wstring.
The automation proxy launcher is only producing ASCII log messages, so
regular strings are fine. In the process, fix an unwanted & in one of
the log messages which causes a format string error.

The video capture test is generating a filename and should be using
FILE_PATH_LITERAL to get the right kind of string for the platform.


Review URL: http://codereview.chromium.org/7519028