Re-write pyauto automation hook WaitForAllDownloadsToComplete.
Now that the Chrome downloads system is more stable since when this hook was originally written, the hook is now updated to work for all downloads (it previously did not work well for dangerous downloads, or downloads from an incognito window).

A new input to the automation hook also requires changes to several pyauto files that invoke the hook.

BUG=69738,91263
TEST=None

Review URL: http://codereview.chromium.org/7544026