Allow sync integration tests to operate on multiple datatypes: Apps, Extensions, Themes
The sync integration tests currently use a class hierarchy where the test classes for each datatype are subclasses of LiveSyncTest. While this design worked in the past, it allows tests to work with only one datatype at a time, and therefore doesn't allow us to test the interplay between datatypes.

This patch is another in the series of patches that will move away from an inheritance model to one where test cases can operate on more than one datatype. It updates the Apps, Extensions and Themes datatypes to the new model, and contains the following changes:

- LiveAppsSyncTest is no longer a class that inherits from LiveSyncTest, but is now a namespace called apps_helper, and contains a bunch of methods that perform various operations related to Apps. Similar changes were made to Extensions and Themes.
- SingleClientLiveAppsSyncTest is renamed to SingleClientAppsSyncTest, and is a subclass of LiveSyncTest. It uses the methods in namespace apps_helper by including its header file and attaching itself to the helper. Similar changes were made to Extensions and Themes.
- LiveSyncExtensionHelper is renamed to SyncExtensionHelper.

BUG=88510
TEST=sync_integration_tests, sync_performance_tests

Review URL: http://codereview.chromium.org/7599019