Revert 97049 - Trying to see if it fixes sync_integration_tests
[Mac] Flip the flag for multi-profiles to on by default.

This also changes ProfileManager::GetProfileInfoCache() to be GetProfileInfo()
and returning the ProfileInfoInterface, making it easier to mock out for testing.
In the few places that require the full ProfileInfoCache, GetMutableProfileInfo()
was added.

Because a handful of Mac unittests make calls into g_browser_process for the
ProfileManager, BrowserTestHelper was expanded to set that up with a
TestingProfile and a FakeProfileInfo so that tests continue to pass.

BUG=60105
TEST=Multi-profiles is on by default.

Review URL: http://codereview.chromium.org/7648037

TBR=rsesek@chromium.org
Review URL: http://codereview.chromium.org/7621031