Multi-Profiles: Add profile avatar if second profile is added
When adding a new profile existing windows wouldn't update to show the profile avatar icon.

The problem was that we were only adding the avatar on window creation.

Fix was to check to see if we should add the avatar when the profile info cache changes.

Note, when adding or removing the avatar we layout the entire window frame. This is more than necessary but it's not a big deal since new profiles are not added very often.

BUG=92740
TEST=


Review URL: http://codereview.chromium.org/7631027