Handle intranet hostnames more consistently.
AutocompleteInput::Parse() now usually marks intranet navigations as UNKNOWN and relies on the HistoryURLProvider to do the right thing.  In turn the HistoryURLProvider will mark URLs it's never seen as high-quality navigations if they have a hostname that's part of a previously-type intranet URL.

So for example, "abc/def" on a clean profile now defaults to searching rather than navigating; but if any URL on host "abc" has been typed before, both "abc/def" and "abc/def ghi" will navigate, despite the space in the second case.

This also changes the default behavior for some similar cases like "abc:81", "user:pass@abc", and "user:pass@abc:81".  Again these will all now default to searching, but navigate if the host in question has been typed.  Unchanged is the behavior of "abc/"; because the scheme-stripping code relies on a single trailing slash triggering a navigation, and because it's convenient for users to force navigation on an unknown host, this is still parsed as a URL.

BUG=6788
TEST=Navigate to "who/", then type "who/a b" and note that the default action is to navigate rather than search.
Review URL: http://codereview.chromium.org/7661028