Tweak typed-count Usage.
Tweak how the typed count is used in the HistoryQuickProvider so that a typed_count of 1 is considered when presenting results.
Adjusted usage of the three factors in RowQualifiesAsSignificant so that the comparison operations are the same.
Adjusted unit tests since the tweak to RowQualifiesAsSignificant changes the expected results just a bit.

BUG=77476
TEST=Adjusted unit tests.
Review URL: http://codereview.chromium.org/7618028