Merge 97248 - Mirror Windows implementation of CancelOnMessageLoopThread.
Before this change, there was a code path where the FilePathWatcher could have been canceled but the delegate_ was not set to NULL. At exit, the destructor would then call Cancel() which would try to post a task to the message loop. With this change, the delegate is correctly set to NULL when the MessageLoop is destroyed even if the FilePathWatcher itself has been canceled. This allows Cancel() from the destructor to do the right thing and not call into the MessageLoop.

This implementation matches the Windows version which does not crash.

BUG=83190
TEST=FilePathWatcherTest.*


Review URL: http://codereview.chromium.org/7655028

TBR=dominich@chromium.org
Review URL: http://codereview.chromium.org/7706005