Do not ignore notifications that come before a LoadComplete.
In the browser, we currently assume that a page has not loaded until a load complete comes from the renderer. However, this does not appear to be the way WebKit behaves. One can move the focus on a webpage that has not completely loaded for example. In such cases, this patch sends a "fake" load complete to keep with our assumptions before allowing the original notification to go through our previous logic.

BUG=90787,93232
TEST=manual (OS X only).


Review URL: http://codereview.chromium.org/7712018