Merge 96634 - Print Preview: Fixing behavior of event listeners.
Now that pending preview requests are being canceled when a new request is issued, the event listeners of the printing controls should not check for |hasPendingPreviewRequest| before issuing new requests, instead they should immediately request a new preview so that the previous one is canceled.

Also all functions that are associated with a specific print request should get the requestID as a parameter, in order to ignore obsolete responses.

I found out that when the print preview tab is reloaded, responses from requests before the reload are reaching the ui code. If reloading the preview tab repeatedly all these request have id=0 and are not being ignored as they should. Therefore I changed the id of the initial request to have a random variable.

BUG=91556, 91769, 91057
TEST=See bug description

Review URL: http://codereview.chromium.org/7550022

TBR=dpapad@chromium.org
Review URL: http://codereview.chromium.org/7744040