Move TooltipWindowGtk from views/widget to ui/base/gtk.
TooltipWidgetGtk is a gtk based helper class that replaces a widget's default tooltip with a RGBA friendly version. It's not closely coupled with views and ui/base/gtk seems to be a better home for it.

This move gets rid of RenderWidgetHostViewGtk's dependency on views.

BUG=none.
TEST=Expect no change. ChromeOS's tooltip should still have rounded corners.

Review URL: http://codereview.chromium.org/7753038