Add invalid rect tracking and repaint scheduling to Layer.
This adds a LayerDelegate interface that can be implemented by an object that wants to know when to repaint.

This is a different model to current layer updating, where code must call SetCanvas when it wishes to push an updated canvas to the layer's texture. 

BUG=none
TEST=see unittest.
Review URL: http://codereview.chromium.org/7822002