Move binding_utils::GetCurrentRenderView() to ExtensionBase and rename ExtensionBase::GetCurrentExtension() and CheckPermissionsForCurrentContext() to make it clear that they actually test the current RenderView.
I'm moving GetCurrentRenderView() because bindings_utils.* is going away in a subsequent change and I needed someplace to put it.

It's OK with me to have it protected because we really shouldn't be using this so much. It is pretty bad to assume anything about a context based on its containing RenderView.

Review URL: http://codereview.chromium.org/7824018