Move the UTF8 text out of .cc files and put them into regular_expressions.txt.
This is temporary.  I think having them in a separate file is too much
maintenance overhead.  So I plan on following up with extending grit to handle
this.  I want to go ahead and get this in to fix people's build and
so the encoding change for pak files can go in.

Review URL: http://codereview.chromium.org/7856001