Data race on boolean in WebDataService::WebDataRequest
Fix for data race in WebDataService::WebDataRequest::Cancel.  Also removes the suppression.  This was happening with very low frequency.  Probably during shutdown.  It appears to have been largely benign because the lock within WebDataService::CancelRequest would protect the important structures.  This change adds a lock to the WebDataRequest::Cancel() operation and makes the query WebDataRequest::IsCancelled() atomic with respect to the consumer and the cancelled state.

BUG=40244
TEST=Run TSAN on bots, see that they stay green.

Review URL: http://codereview.chromium.org/7870007