[Mac] Break stale controller refs in infobar close.
Infobars can be closed due to a navigation while the user is
manipulating controls on the infobar.  This can result in Objective-C
messages to a deallocated controller.  If the controller is still
live, the action could attempt to dereference a NULL delegate.  This
attempts to knock down both cases.

BUG=94639, 92381
TEST=See bug, and monitor crash server.


Review URL: http://codereview.chromium.org/7876004