touchui: Fix the maximization of a NativeWidgetGtk.
Using gtk_widget_set_size_request leaves the window in a very weird state: it
changes the apparent size, but if the backing X window was larger before the
request, then it doesn't get resized to the smaller size. So subsequent calls to
get the bounds of the widget gets the pre-maximized larger size. So use the
proper API (gtk_window_resize) to do the window resizing.

BUG=none
TEST=manually

Review URL: http://codereview.chromium.org/7887019