Stop restricting the sender object path for incoming signals.
Restricting the sender path makes it impossible to test
signal handling with dbus-send --type=signal, as it uses
"/" as the sender object path.

For now, let's remove the restriction. We can make this restriction
customizable when it becomes necessary.

This is for http://codereview.chromium.org/7862020/

TEST=dbus_unittests
BUG=none

Review URL: http://codereview.chromium.org/7845029