[Sync] Avoid passing around EntryKernelMutationSet by value
Change EntryKernelMutationSet to EntryKernelMutationMap so that it can be filled in-place.

Create a new class SharedEntryKernelMutationMap which is a thread-safe
immutable wrapper around an EntryKernelMutationMap.  Pass it around instead of the mutation map directly.

Generalize SharedValue and move it to sync/util.

Move the mutation limit to the JsTransactionObserver.

BUG=96483
TEST=

Review URL: http://codereview.chromium.org/7889020