Removing singelton property of MediaStreamManager.
This is almost the same patch [1] as was reverted last week due to flakiness in one of the Win tests, but for one change: MediaStreamManager is not called from MediaStreamDispatcherHost::OnChannelClosing, to avoid shut-down crash. This will be changed as soon as the shut-down issue is solved.

[1] http://codereview.chromium.org/7649016/

Review URL: http://codereview.chromium.org/7948004