Remove "official extras" archive files that don't exist.
AFAICT, these don't exist on any current branch and haven't for a while. This
fixes the new Windows stage_build script which considers missing files an error
(as it always should have, but since we used the same scripts for all channels,
we had to assume missing files might rightly not exist for some channels).

Eventually all the "official extras" should be taken out of the python scripts
and moved to FILES.cfg, but due to the Code Yellow, I'm just cleaning them up
in-place, as the safest, backwards-compatible solution.
Review URL: http://codereview.chromium.org/8023002