Use precompiled headers on Windows only when a flag is set.
I would have preferred if precompiled headers could be enabled by
default. The motivation for this change is primarily that Visual
Studio does the wrong thing on preprocessor flag changes - instead of
rebuilding precompiled headers, it errors out on any file that uses
them, saying that the precompiled header was built with different
flags! This causes a world of hurt on our infrastructure stuff,
particularly on trybots as they may be jumping back and forth between
revisions before and after a flag change.

Prior to this change, only Debug builds used precompiled headers and
Release builds did not. The reason Release builds did not was that
official builds were running out of memory with precompiled headers
enabled. The distinction of Debug vs. Release is no longer necessary
after the flag is added, so removing that extra bit of complexity from
the .gyp files.

BUG=none
TEST=it builds


Review URL: http://codereview.chromium.org/8041022