Move plugin loading out of process on Mac and Linux.
This creates a new set of IPC messages for the utility process to load plugins
to get the WebPluginInfo data in a separate process. Previously this was done
in the browser process, but that involves loading arbitrary third-party code
into the address space and then executing it.

BUG=17863,95114
TEST=Plugins work as before.

Review URL: http://codereview.chromium.org/7889025