Fixes painting issue with bookmarks bar
When the bookmarks bar was showing, opening a dialog which paints on top of the bookmark bar in touch_ui
and dismissing it would leave ghosts where the dialog used to be.
The reason is that the root view has a layer. It would paint the entire view tree into the root view including content which is contained by child layers

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/8046019