Add a content_shell specific DownloadManagerDelegate.
Also fix a race condition that existed in resource dispatching but was never tickled by Chrome. The problem is if in BufferedResourceHandler::OnReadCompleted(), CompleteResponseStarted calls the next handler's OnResponseStarted. If that pauses the request, then the next handler's OnReadCompleted() would still get called which is not correct.

BUG=90445
Review URL: http://codereview.chromium.org/8065019