Clean up WebUI task manager CSS and disable selection.
This disables user selections and uses the default cursor in
the task manager dialog; previously, the title and the
headings were selectable (sadly, the possibly-useful-to-copy
data being displayed in the dialog was already
unselectable).

I'm also sorting the CSS rules as instructed by the style
guide.

BUG=chromium-os:20738
TEST=manual

Review URL: http://codereview.chromium.org/8089011