Avoid a possible race in sync integration tests
I haven't been able to confirm this, but I suspect there are cases where
IsSynced could be evaluated before a transaction on the syncer thread and
IsTypeEncrypted would be evaluated afterwards.  If that did happen, we could
leave WaitForTypeEncryption long before we ought to, which is likely to result
in test failures.

Reordering the expressions doesn't really address the root cause of the issue.
However, we know that these conditions only change in certain ways, so the
reordering should effectively hide the bug.

BUG=98607
TEST=sync_integration

Review URL: http://codereview.chromium.org/8086010