Revert 103602 - Fix the problem that sometimes panel on Windows is not set to top-most. It borked Linux Touch and Linux Views compile
This could happen when we try to make the window top-most after it has been
created without activation. This seems to be a bug or feature on Windows that
some users have been complaining about.

To fix this problem, I changed BrowserFrame::InitBrowserFrame to pass the
keep_on_top flag when the panel is to be created.

I have also added windows specific testing code to check the styles in the
browser test.

BUG=none
TEST=browser test
Review URL: http://codereview.chromium.org/8098007

TBR=jianli@chromium.org
Review URL: http://codereview.chromium.org/8103017