Create scanning script to determine header dependencies.
This python script will take a set of input files and header include directories and scan through the sources trying to find all header references to generate a list of file dependencies.  The scanner does not use a preprocessor so all headers in the source even if conditionally compiled out, are returned.  Files which can not be found in the file system are ignored.


BUG= http://code.google.com/p/chromium/issues/detail?id=96782
TEST= manual
Review URL: http://codereview.chromium.org/8037013