chromeos: Use the new LockEnabled flimflam device property.
Now that there is a property that explicitly indicates whether
a SIM PIN lock is enabled, use that instead of a heuristic to
decide what to display in the "Lock SIM card" checkbox in the
internet options Security tab. This means that the display will
be correct even across a logout/login cycle or a Chrome crash.

BUG=chromium-os:14253
TEST=Boot the system while SIM PIN locking is enabled. Enter the
correct PIN to unlock the SIM. Log out and then log back in, causing
Chrome to restart. Go back to the Security tab, and observe that
the "Lock SIM card" checkbox is checked. I also exercised the other
PIN operations to ensure that everything else works as it did before.

R=nkostylev@chromium.org

Review URL: http://codereview.chromium.org/8124007