Revert 101893 "Make IsSynced() false when syncing in progress"
When trying to fix the "commit error" (issue 94670), I found that the
integration tests were failing because the test harness was returning early
from AwaitMutualSyncCycleComplete (issue 95742).  I tried to fix this by adding
a check of AllStatus' summary flag in the IsSynced() condition (r101893).  It
turns out that this was a bad idea, because this flag doesn't do a very good
job of reflecting what's actually happening (issue 98346). 

This commit reverts the change that "fixed" issue 95742.  We will need to
find some other way to prevent the harness from exiting early.

BUG=95742,98246,94670
TEST=sync_integration


Review URL: http://codereview.chromium.org/8114016