PPAPI Fullscreen: Hack plugin's attributes to force resizing of the plugin to fill in the screen in SetFullscreen. Reset plugin size in ViewChanged when we exit. Update tests.
We are forced to resize the plugin in the Pepper layer because WebKit does yet do it.
It only moves the plugin to the middle of the screen with the expectations the
web page will contain the following css:
   :-webkit-full-screen {
        width: 100%; height: 100%;
    }
That's not acceptable for Pepper/NaCl users.

BUG=41780,98474
TEST=in CL
Review URL: http://codereview.chromium.org/8228023