gtk: fix bulge in custom frame
problem: when there is no Other Bookmarks button showing (because the folder is empty), you can see a small beige bulge in the custom frame at the level of the bookmark bar
cause: the other_padding GtkAlignment is still showing although the button itself is hidden (I'm not sure why this alignment extends outside the bounds of the bookmark bar, but it does)
solution: show/hide that widget instead of the button widget.

problem 2: the Other Bookmarks button is not shown/hidden when elements are added to or removed from it.
solution: call UpdateOtherBookmarksVisibility more aggressively

BUG=100724
TEST=manual


Review URL: http://codereview.chromium.org/8336011