QuotaManager: add a method to delete host data.
Added QuotaManager::DeleteHostData method that takes host name
and deletes data for all origins related to the host.

As it is planned to provide quota manager users with a list
of hosts, they would not be able to delete host data, as
currently it is only possible to delete data by origin.

R=kinuko@chromium.org
BUG=none
TEST=QuotaManagerTest.DeleteHostData*

Review URL: http://codereview.chromium.org/8349032