aura: Improve initial host window size.
This adds a --aura-host-window-size switch (taking an
argument of the form "1024x768") that can be used to set the
initial size for the host window.

It also makes the Linux implementation automatically size
the window to cover the entire screen if an X window manager
isn't present (i.e. we're running on a Chrome OS device).

BUG=100979
TEST=tried it on a chrome os device

Review URL: http://codereview.chromium.org/8362002