Merge 106796 - Handle the change from CaretBounds to SelectionBounds for PPAPI Plugins.
BUG=101173
TEST=Manual: run ppapi_example_ime and verify the candidate window to pop up
in a correct place, or open a Flash website with text field on Chrome OS, and
verify that the candidate window is displayed in the bottom-left corner of
the Flash rect.

This CL is to reflect the refactoring http://crrev.com/105699 on IPCs also
for Pepper plugins, and make IME candidate windows to be displayed on intended
positions.


Review URL: http://codereview.chromium.org/8363015

TBR=kinaba@chromium.org
Review URL: http://codereview.chromium.org/8382011