Fix about:gpu's wrong status of accelerated-canvas-2d.
The decision whether to support accelerated-canvas-2d is made on the renderer side.  This CL duplicates the same logic on the browser side, so about:gpu reports the correct feature status.

An alternative is to add an IPC messege to query renderer side about the feature status, which isn't very appealing.

BUG=100722
TEST=about:gpu
Review URL: http://codereview.chromium.org/8341036