Merge 107707 - Exempt component and policy-installed extensions from policy blacklist check.
Component extensions and policy-installed extensions shouldn't be affected by the policy blacklist: The former are considered internal UI that happens to be part of an extension, and they should have first-class policy knobs if we want to disable them. For the latter, the admin specifying them is enough indication that they should be installed, and requiring them to explicitly white-listed in some cases is redundant.

BUG=chromium-os:16702, chromium-os:21281
TEST=On ChromeOS, configure ExtensionInstallBlacklist = * and load chrome://bookmarks


Review URL: http://codereview.chromium.org/8380006

TBR=mnissler@chromium.org
Review URL: http://codereview.chromium.org/8417022