Merge trunk r100149 to the 15.0.874 branch.
Disable PIE for dump_syms. The framework is too big to reliably be able to
mmap it in a 32-bit process with PIE on. There may not be a big enough gap in
the address space to mmap it.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/7840040

TBR=mark@chromium.org
Review URL: http://codereview.chromium.org/8413045