Allow gerrit to send login credentials with request for json build status.
This allows the listed domains to make XHR requests to the status server along
with login credential cookies. The withCredentials option cannot be used with
the wildcard Access-Control-Allow-Origin, so we introduce a new parameter that
the listed domains can use to request that the server set
Access-Control-Allow-Credentials.

BUG=None
TEST=None
Review URL: http://codereview.chromium.org/8341104