Merge 107841 - Fix a crash in FullscreenExitBubbleController when the user clicks the "Exit full screen" button.
The crash was due to the fact that Browser::OnDenyFullscreenPermission causes BrowserWindow::ExitFullscreen to be called, which causes the FullscreenExitBubbleController to be destroyed. [FullscreenExitBubbleController deny:] was sending a -hideSoon message to its zombie self following the call to Browser::OnDenyFullscreenPermission.

BUG=101835
TEST=The browser doesn't crash when you click "Exit full screen" on a fullscreened page.


Review URL: http://codereview.chromium.org/8399017

TBR=jeremya@chromium.org
Review URL: http://codereview.chromium.org/8351086