Fixed a bug where Linux users couldn't download to NTFS file systems.
The problem is that chmod() fails on NTFS and FAT file systems.
That should not cause the download to fail.

The bug was introduced in r99579 (Chrome 15).  I tested it with the previous revision and that one, to verify that as the point where the bug first showed up.

BUG=50104,102200
TEST=Download a file on Ubuntu, to an NTFS disk.


Review URL: http://codereview.chromium.org/8438048