Fix crash in PSS:SetEncryptEverything() and in TypedUrlModelAssociator
Changed the SetEncryptEverything() signature so it doesn't take any parameters,
and is only called when we are trying to encrypt everything.

Got rid of an overzealous DCHECK() in TypedUrlModelAssociator::WriteToHistoryBackend()
which was incorrectly assuming that all items in the history DB must have at
least one visit (this is not true in the case of bookmarks). Added a test for
this case.

BUG=102552
TEST=Sync with an account that has both a bookmark and typed visit for a URL.


Review URL: http://codereview.chromium.org/8412006