Fix css so that incognito warning doesn't overlap next line.
When the incognito warning is shown, its max height was being increased
from 0 to 100px in order to animate it opening. However, 100px is
insufficient because with a narrow window width and/or large fonts
(300% page zoom is the max), the height of the warning is more than 100px.
Changed to 1000px which more than sufficient for the narrowest window and
maximum page zoom. The actual height will be only what is needed, so there's
no harm in making this too large.

BUG=100767
TEST=Show the incognito warning in extension settings with a narrow window.

Review URL: http://codereview.chromium.org/8480002