Avoid blocking round-trips for resizing GLX image transport surface
In the GLX case, when resizing the image transport surface, we immediately have a usable destination surface, without having to wait for an ACK from the browser. However, we don't want to release the old surface (that is still usable) until we have the new bits, so delay the notification until the swap.

BUG=58782
TEST=Aura browser on stumpy, observe that resizing frames leads to content being updated about once every 3 UI frames instead of about 6.


Review URL: http://codereview.chromium.org/8490028