Revert 109357 - Support tracking of IPC messages as tasks in profiler
This was previously landed as 108752, but reverted in 109004
by Joi to try to stop flaky failures in ASAN bot when
running browser_tests..
The failures included:
IncognitoSplitMode
TwoExtensionsCrashBothAtOnce
TwoExtensionsOneByOne
AutoUpdate
CognitoNoFile
IncognitoNoFile
OnChangedNotificationsBetweenBackgroundPages
SplitModeIncognito

I'm now landing 1 hook at a time back into the build, to
see which one(s) tickle flakiness.
The other hooks are commented out, so that I can
try several alternatives and figure out what is
tickling the problem.

The restructuring of the code was already landed, and did
not *appear* to induce the failures.  Hence it is likely
that one of the IPC hooks is the problem.

If I need to, I'll TBR various owners, since I've already gotten permission for this landing.


r=rtenneti
Review URL: http://codereview.chromium.org/8499039

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/8506037