Revert 109960 - Send WM_DEVICECHANGE message through SystemMonitor
WM_DEVICECHANGE is sent when there's been a change to devices or the computer;
specifically when a USB device is connected or disconnected. This is intended
for use in support of Gamepads for more performant polling and
connect/disconnect testing. Currently only on Windows, though seems reasonable
to add for other platforms in the future.

BUG=79050


Review URL: http://codereview.chromium.org/8523021

TBR=scottmg@chromium.org
Review URL: http://codereview.chromium.org/8528044