Reduce race potential for lazy testing initialization of profiler
Initialization of the profiler is supposed to happen while
single threaded.  The tests however often skip affirmative
initialization, and rely on lazy initialization.  This
CL tightens up that process, to reduce changes of races
(even benign races) during testing.

I also added a suppression for examination of the state,
which is always a benign race.  At initialization time
we rely on a lock (now) to ensure race freedom (if we
get a racy value of state).

r=rtenneti
Review URL: http://codereview.chromium.org/8570007