Enable content component (again!).
This version fixes a problem with the windows incremental linking build. We can't incrementally link chrome when content is being built as a DLL because chrome links in webkit_glue and webkit_glue depends on symbols defined in content. We can remove this when we fix glue.

R=jam@chromium.org
BUG=90442, 98755.

Review URL: http://codereview.chromium.org/8570024