GTK: Always relayout the tab strip when the tab strip bounds change.
TabStripGtk::OnSizeAllocate() could fail to call Layout() or schedule Layout()
to be called in the future (and was reusing the code to resize the internal
tabs when a single tab was closed). The views implementation of the tabstrip
didn't use that path and would always relayout the entire tabstrip on bounds
changes.

BUG=79094
TEST=Create a window with more than 20 tabs in it. Very slowly shrink the window, one pixel at a time. The min/max/close buttons shouldn't overlap with the new tab button.


Review URL: http://codereview.chromium.org/8568026