Fix bug preventing pyauto suite CHROMEOS_PERF from running on ChromeOS.
A string class variable in netflix.py is named starting with the substring
'test', which is being interpreted as a test case when pauto is invoked with
suite CHROMEOS_PERF (since that suite runs perf.py on ChromeOS, which in turn
imports from netflix.py).  This CL changes the name of that class variable
to avoid this problem, and also includes a few minor, related edits.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/8677013