Check that thread contexts are cleaned up during profiling
There was a Thread Local Store regression, wherein
thread cleanup notifcation was not sent to TLS
users.  This CL ensures that when this happens,
that Chrome will crash.

For now, I'm landed with a short-circuit (returns true
all the time) in the test.  I'll remove the TODO when
we have XP properly supported.

r=rtenneti
BUG=103209
Review URL: http://codereview.chromium.org/8606001