Controlled settings indicator and bubble for chrome://preferences.
The idea is that every setting that is under control by an external entity (e.g. policy, extension) will be annotated with an icon that, when click on, will show a bubble with more information. This implements the indicator icon and bubble. Going through the prefs UI and inserting the thing at the appropriate places is left for subsequent CLs :)

BUG=chromium:104955
TEST=No observable changes.

Review URL: http://codereview.chromium.org/8480037