Get more information about types of images being downloaded.
This adds a new enum Download.ContentImageType rather than adding additional entries to Download.ContentType.

The downside is that there is a bit more code. The upside is that we don't have to worry about the image entries getting fragmented across the Download.ContentType enum, and there is a top-level entry for images in the Download.ContentType histogram to compare against other general media types.

TEST=Download a few images, make sure Download.ContentType and Download.ContentImageType look sane.


Review URL: http://codereview.chromium.org/8734006