Prerenders that send a ViewHostMsg_OpenURL are canceled with FINAL_STATUS_OPEN_URL.
This message is sent when a navigation is initiated from a renderer but requires browser intervention, typically because a new render process will be created. Examples of where this happens are redirects from a "normal page" to a hosted app domain, or transitions from one origin to another when --enable-strict-site-isolation is specified.

BUG=105972
TEST=Install YouTube app. Start Chrome with --prerender=enabled. Go to prerender-test.appspot.com, and prerender youtube.com. Click on it and www.youtube.com should appear. Check chrome://net-internals#prerender and make sure that the youtube prerender was canceled with OpenURL as the final status.

Review URL: http://codereview.chromium.org/8785007