Add some strings for disabling app notifications.
We initially went with a "disable notifications" menu item, but recently talked
about inverting this and having "enable notifications" or "allow notifications"
instead. I'm checking these in separately to make the translation deadline
so we have the option to change before the M17 release.

BUG=106268
TEST=none

TBR=finnur@chromium.org

Review URL: http://codereview.chromium.org/8801032