Win AcceleratedSurface no longer polls queries on separate thread.
This is basically a reversion of the final patch set (#4) of this:
http://codereview.chromium.org/8395012

The reason is because tasks posted to the query thread can extend the lifetime of the D3D device such that it is destroyed on the query thread. The D3D thread must be destroyed on the thread it was created on though.

The suggestion was a good one and I might revisit it in the future.

I also added a missing lock and switched it to load the D3D9 DLL rather than assume it is already loaded by something else.
Review URL: http://codereview.chromium.org/8867005