Fix webgl conformance test list generation script.
The current code is for update time run, so it can generate correct list per OS.  However, we changed it to run locally and check the generated list in, so we need to have OS conditions for tests.

Also, add DEBUG/RELEASE modifier so we can mark test failures better.

BUG=
TEST=gpu_tests built
R=kbr
Review URL: http://codereview.chromium.org/8885003