Add utility function to wrap a paragraph of text.
This is needed for my upcoming change to use this functionality
in a new CanvasSkia text drawing implementation.

This is exposed as a standalone function in text_elider.h but
implemented using an internal helper class in text_elider.cc.
This is similar to how the existing |ElideRectangleString()|
is implemented.

The new function provides four modes of dealing with long words
when wrapping text which correspond to the different ways the
current implementation of |CanvasSkia::DrawStringInt()| can wrap
text, based on the input flags.

BUG=105550
TEST=New unit tests added to text_elider_unittest.cc

Review URL: http://codereview.chromium.org/8234003