Merge 114685 (to branch 972) - Avoid including gtest file directly, and go through base instead
We should consistently:
#include "base/gtest_prod_util.h"
which does the gtest include (from the right place, in a
second repo, hidden in our tree).


This is critical to being able to supplant the FRIEND_TEST
macro, which we're experimenting with to optimize builds.

r=brettw
TBR=other owners (this is a trivial change)
BUG=107727
Review URL: http://codereview.chromium.org/8954017

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/8963009