Don't tell browser to swap until swap actually executed.
By deferring to fence, we're ensuring that the GL commands are executed GPU side before sending a message to the browser to act on them. This is lighter weight than a glFinish because the other commands buffers on the channel (e.g. WebGL) can continue executing while we're waiting for the fence.

TEST=by hand building use_aura running http://webglsamples.googlecode.com/hg/aquarium/aquarium.html

Review URL: http://codereview.chromium.org/8497015