Add ability to pass arguments to browser_tests Add the ability to pass arbitrary arguments to the browser_tests. This is needed for diagnosing and debugging the tests on a CrOS device/VM.
To use, add pass_to_tests=argument to the --args parameter, so for example, to run the browser_tests in single process mode you would use,
./run_remote_tests.sh --use_emerged --remote localhost --ssh_port 9222 desktopui_BrowserTest.control.custom --args 'gtest_filter=TestSuite.Test pass_to_tests=single_process'

The -- will automatically be added to the argument. Multiple pass_to_tests parameters can be specified in any order; all will be processed and the arguments added to the command line for the tests.


R=zelidrag@chromium.org,stevenjb@chromium.org,nirnimesh@chromium.org
BUG=None
TEST=Tested the new pass_to_tests= argument and verified existing arguments still work correctly.


Review URL: http://codereview.chromium.org/8992025